### 2025-05-02 #8153 @editorialbot
[PRE REVIEW]: simulist: An R package to simulate disease outbreak line list and contacts data
**Submitting author:** @joshwlambert (<a href="http://orcid.org/0000-0001-5218-3046">Joshua Lambert</a>)
**Repository:** https://github.com/epiverse-trace/simulist
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.0
