### 2025-05-02 8677fdd
improve @return documentation for exported functions, closes #234
