### 2025-05-02 d5f7ee1
improve @return documentation for exported functions, closes #234
