### 2025-05-02 0973718
bump R minimum version to v4.2 due to dependency
