### 2025-05-02 e3eac85
bump R minimum version to v4.2 due to dependency
