+++ DESCRIPTION
@@ -29,7 +29,7 @@ URL: https://github.com/epiverse-trace/simulist,
     https://epiverse-trace.github.io/simulist/
 BugReports: https://github.com/epiverse-trace/simulist/issues
 Depends: 
-    R (>= 4.1.0)
+    R (>= 4.2.0)
 Imports:
     checkmate,
     english,
