### 2025-05-06 407b489
Update `CITATION.cff`
