### 2025-07-11 #8153 @fabian-s
:wave: @seabbs @LogIN- @JDRomano2
would any of you be willing to review this submission for JOSS?

*Please be so kind to let me know if you're not interested, so that I can start looking for other reviewers.*
