### 2025-07-16 fe29394
add censor_linelist function, relates #207
