+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(censor_linelist)
 export(create_config)
 export(messy_linelist)
 export(sim_contacts)
