### 2025-07-16 f0dfd2f
add censor_linelist function, relates #207
