### 2025-07-21 d1ae7d6
add weekend effects to censor_linelist, relates #207
