### 2025-07-21 682d102
add weekend effects to censor_linelist, relates #207
