### 2025-07-21 655c9ff
add unit test for weekend_effects in censor_linelist
