+++ tests/testthat/test-censor_linelist.R
@@ -94,6 +94,18 @@ test_that("censor_linelist works for integer interval censoring", {
   expect_true(all(date_col_int))
 })
 
+test_that("censor_linelist works for weekend_effects", {
+  ll_cens <- censor_linelist(
+    linelist = ll,
+    interval = "daily",
+    reporting_artefact = "weekend_effects"
+  )
+  expect_true(all(c("Saturday", "Sunday") %in% weekdays(ll$date_reporting)))
+  expect_false(
+    any(c("Saturday", "Sunday") %in% weekdays(ll_cens$date_reporting))
+  )
+})
+
 test_that("censor_linelist fails correctly for invalid input", {
   expect_error(
     censor_linelist(linelist = ll, interval = 1.5),
