### 2025-07-21 caffdf1
apply weekend effect for numeric and character intervals in censor_linelist
