### 2025-07-21 3f1b6f7
add unit test for weekend_effect for period in censor_linelist
