### 2025-07-28 #8153 @fabian-s
> 👋 [@seabbs](https://github.com/seabbs) [@LogIN-](https://github.com/LogIN-) [@JDRomano2](https://github.com/JDRomano2) would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. _Please be so kind to let me know if you're not interested, so that I can start looking for other reviewers._
