### 2025-08-11 #8153 @joshwlambert
Sorry for the late reply, I've been on leave for the past couple of weeks. From the JOSS reviewer management system I've found a few people that have listed _R_ in languages and _epidemiology_ in topic areas: @ruarai, @nilabhrardas,
@mattwarkentin, @epiben.
