### 2025-08-18 0110c4d
remove test for period weekend_effects
