### 2025-08-18 9da20ec
add offset argument to censor_linelist
