### 2025-08-18 6798ff1
add offset argument to censor_linelist
