### 2025-08-18 4ce504a
add censoring to README
