+++ README.md
@@ -55,6 +55,8 @@ Post-process simulated line list data for:
 
 :date: Real-time outbreak snapshots with right-truncation <br> :memo:
 Messy data with inconsistencies, mistakes and missing values <br>
+:ledger: Censor dates to daily, epi- and iso-weekly, yearly and other
+groupings <br>
 
 ## Installation
 
