### 2025-08-18 2fd2b5e
remove messy_linelist unit test for multiple type vector (R coerces to same type)
