### 2025-08-20 2125378
update sim_contacts $was_case column to boolean, fixes #242
