### 2025-08-20 e351155
update sim_contacts $was_case column to boolean, fixes #242
