### 2025-08-20 74e32e9
update snapshot tests for $was_case change, relates #242
