+++ tests/testthat/_snaps/sim_outbreak.md
@@ -811,332 +811,332 @@
       325               John Trujillo                Haley Hamlin  82   f
       326                Haley Hamlin             Larissa Herrera  30   f
           date_first_contact date_last_contact was_case           status
-      1           2022-12-29        2023-01-03        Y             case
-      2           2022-12-28        2023-01-01        Y             case
-      3           2022-12-26        2023-01-02        N lost_to_followup
-      4           2022-12-28        2023-01-04        Y             case
-      5           2022-12-31        2023-01-04        Y             case
-      6           2022-12-31        2023-01-06        Y             case
-      7           2023-01-01        2023-01-02        N   under_followup
-      8           2023-01-03        2023-01-09        N   under_followup
-      9           2023-01-02        2023-01-07        Y             case
-      10          2023-01-03        2023-01-07        Y             case
-      11          2023-01-03        2023-01-11        Y             case
-      12          2022-12-30        2023-01-08        N   under_followup
-      13          2023-01-04        2023-01-09        Y             case
-      14          2023-01-09        2023-01-17        N   under_followup
-      15          2023-01-07        2023-01-16        N   under_followup
-      16          2023-01-11        2023-01-13        N   under_followup
-      17          2023-01-09        2023-01-17        Y             case
-      18          2023-01-07        2023-01-14        Y             case
-      19          2023-01-08        2023-01-13        N   under_followup
-      20          2023-01-09        2023-01-15        N   under_followup
-      21          2023-01-07        2023-01-17        Y             case
-      22          2023-01-12        2023-01-16        N   under_followup
-      23          2023-01-08        2023-01-15        Y             case
-      24          2023-01-11        2023-01-14        N   under_followup
-      25          2023-01-08        2023-01-13        N lost_to_followup
-      26          2023-01-09        2023-01-16        Y             case
-      27          2023-01-07        2023-01-15        N   under_followup
-      28          2023-01-08        2023-01-14        Y             case
-      29          2023-01-07        2023-01-13        Y             case
-      30          2023-01-10        2023-01-16        Y             case
-      31          2023-01-10        2023-01-17        N   under_followup
-      32          2023-01-12        2023-01-15        Y             case
-      33          2023-01-12        2023-01-18        Y             case
-      34          2023-01-16        2023-01-21        Y             case
-      35          2023-01-20        2023-01-23        N lost_to_followup
-      36          2023-01-19        2023-01-25        N lost_to_followup
-      37          2023-01-19        2023-01-23        Y             case
-      38          2023-01-18        2023-01-24        Y             case
-      39          2023-01-20        2023-01-26        N lost_to_followup
-      40          2023-01-17        2023-01-21        Y             case
-      41          2023-01-15        2023-01-20        N   under_followup
-      42          2023-01-14        2023-01-21        N   under_followup
-      43          2023-01-18        2023-01-21        Y             case
-      44          2023-01-19        2023-01-25        Y             case
-      45          2023-01-12        2023-01-18        N   under_followup
-      46          2023-01-23        2023-01-28        Y             case
-      47          2023-01-26        2023-01-29        N   under_followup
-      48          2023-01-19        2023-01-28        N lost_to_followup
-      49          2023-01-20        2023-01-28        Y             case
-      50          2023-01-22        2023-01-25        Y             case
-      51          2023-01-22        2023-01-26        N   under_followup
-      52          2023-01-20        2023-01-27        Y             case
-      53          2023-01-23        2023-01-29        Y             case
-      54          2023-01-17        2023-01-24        Y             case
-      55          2023-01-21        2023-01-26        Y             case
-      56          2023-01-18        2023-01-26        N   under_followup
[…]
