+++ tests/testthat/_snaps/sim_contacts.md
@@ -25,26 +25,26 @@
       19          Dominica Abeyta        Arkaan al-Faris  22   m         2023-02-04
       20          Dominica Abeyta          Olivia Carson  70   f         2023-02-03
          date_last_contact was_case           status
-      1         2023-01-04        N   under_followup
-      2         2023-01-06        Y             case
-      3         2023-01-18        Y             case
-      4         2023-01-18        N lost_to_followup
-      5         2023-01-17        N   under_followup
-      6         2023-01-18        Y             case
-      7         2023-01-20        N lost_to_followup
-      8         2023-01-19        Y             case
-      9         2023-01-30        N   under_followup
-      10        2023-01-22        Y             case
-      11        2023-01-22        N   under_followup
-      12        2023-01-31        Y             case
-      13        2023-01-26        N   under_followup
-      14        2023-02-03        Y             case
-      15        2023-02-04        N lost_to_followup
-      16        2023-02-06        Y             case
-      17        2023-02-09        N   under_followup
-      18        2023-02-08        N   under_followup
-      19        2023-02-12        Y             case
-      20        2023-02-09        N lost_to_followup
+      1         2023-01-04    FALSE   under_followup
+      2         2023-01-06     TRUE             case
+      3         2023-01-18     TRUE             case
+      4         2023-01-18    FALSE lost_to_followup
+      5         2023-01-17    FALSE   under_followup
+      6         2023-01-18     TRUE             case
+      7         2023-01-20    FALSE lost_to_followup
+      8         2023-01-19     TRUE             case
+      9         2023-01-30    FALSE   under_followup
+      10        2023-01-22     TRUE             case
+      11        2023-01-22    FALSE   under_followup
+      12        2023-01-31     TRUE             case
+      13        2023-01-26    FALSE   under_followup
+      14        2023-02-03     TRUE             case
+      15        2023-02-04    FALSE lost_to_followup
+      16        2023-02-06     TRUE             case
+      17        2023-02-09    FALSE   under_followup
+      18        2023-02-08    FALSE   under_followup
+      19        2023-02-12     TRUE             case
+      20        2023-02-09    FALSE lost_to_followup
 
 # sim_contacts works as expected
 
@@ -79,31 +79,31 @@
       24    Shan Saldanha           Preston Miles  75   m         2023-01-03
       25  Alexander Brown              Drue White  31   m         2022-12-30
          date_last_contact was_case           status
-      1         2023-01-04        N   under_followup
-      2         2023-01-04        Y             case
-      3         2023-01-10        Y             case
-      4         2023-01-05        N   under_followup
-      5         2023-01-05        N   under_followup
-      6         2023-01-03        N          unknown
-      7         2023-01-06        Y             case
-      8         2023-01-05        Y             case
-      9         2023-01-05        N lost_to_followup
-      10        2023-01-04        Y             case
-      11        2023-01-04        N   under_followup
-      12        2023-01-05        Y             case
-      13        2023-01-05        N   under_followup
-      14        2023-01-03        Y             case
-      15        2023-01-02        Y             case
-      16        2023-01-05        Y             case
-      17        2023-01-09        N   under_followup
-      18        2023-01-06        N   under_followup
-      19        2023-01-06        N          unknown
-      20        2023-01-06        N lost_to_followup
-      21        2023-01-10        Y             case
-      22        2023-01-06        N   under_followup
-      23        2023-01-07        N   under_followup
-      24        2023-01-06        Y             case
-      25        2023-01-05        N   under_followup
+      1         2023-01-04    FALSE   under_followup
+      2         2023-01-04     TRUE             case
[…]
