### 2025-08-20 b3a8ce5
update subsetting using $was_case in vis-linelist vignette
