### 2025-08-20 90a3165
Update `CITATION.cff`
