### 2025-08-20 ade8c01
Update `CITATION.cff`
