+++ CITATION.cff
@@ -22,7 +22,7 @@ authors:
   given-names: Joshua W.
   email: joshua.lambert@lshtm.ac.uk
   orcid: https://orcid.org/0000-0001-5218-3046
-- family-names: Tamayo
+- family-names: Tamayo Cuartero
   given-names: Carmen
   email: carmen.tamayo-cuartero@lshtm.ac.uk
   orcid: https://orcid.org/0000-0003-4184-2864
@@ -34,7 +34,7 @@ preferred-citation:
     given-names: Joshua W.
     email: joshua.lambert@lshtm.ac.uk
     orcid: https://orcid.org/0000-0001-5218-3046
-  - family-names: Tamayo
+  - family-names: Tamayo Cuartero
     given-names: Carmen
     email: carmen.tamayo-cuartero@lshtm.ac.uk
     orcid: https://orcid.org/0000-0003-4184-2864
