### 2025-08-20 6895b2b
add NEWS items for v0.6.0 release
