### 2025-08-28 #8781 @joshwlambert
@fabian-s I've reviewed the bibliography and the DOI check, the missing DOI is not an official DOI for the R language, but rather for its help pages. Therefore I'm happy that all the DOIs that are available for references are in `paper.bib`.
