### 2025-08-28 06bf54f
Update `CITATION.cff`
