+++ dev/pkgdown.yml
@@ -1,16 +1,0 @@
-pandoc: 3.1.11
-pkgdown: 2.1.3
-pkgdown_sha: ~
-articles:
-  age-strat-risks: age-strat-risks.html
-  age-struct-pop: age-struct-pop.html
-  design-principles: design-principles.html
-  reporting-delays-truncation: reporting-delays-truncation.html
-  simulist: simulist.html
-  time-varying-cfr: time-varying-cfr.html
-  vis-linelist: vis-linelist.html
-  wrangling-linelist: wrangling-linelist.html
-last_built: 2025-08-20T16:32Z
-urls:
-  reference: https://epiverse-trace.github.io/simulist/reference
-  article: https://epiverse-trace.github.io/simulist/articles
