### 2025-10-15 #8781 @fabian-s
@epiben please let us know if you are still willing to review this paper for JOSS and when we can expect your review.
