### 2025-11-10 d49a6bd
update create_config test expectations with new list element
