+++ tests/testthat/_snaps/sim_contacts.md
@@ -3,27 +3,27 @@
     Code
       sim_contacts()
     Output
-                             from                     to age sex date_first_contact
-      1  Louis Baltazar-Francisco        Masooma Sheldon  58   f         2022-12-27
-      2  Louis Baltazar-Francisco      Rachel Fredericks  29   f         2022-12-30
-      3         Rachel Fredericks          Makiah Thomas  24   f         2023-01-12
-      4         Rachel Fredericks         Mersadez Swift  42   f         2023-01-11
-      5         Rachel Fredericks      Mitchell Sandoval  48   m         2023-01-09
-      6             Makiah Thomas       Saahira el-Laham  76   f         2023-01-11
-      7             Makiah Thomas           Celsa Bellow  39   f         2023-01-14
-      8             Makiah Thomas Stephanie Garcia Mayen  24   f         2023-01-12
-      9          Saahira el-Laham        Laurel Lehmkuhl  53   f         2023-01-20
-      10   Stephanie Garcia Mayen           Cade Amerson  86   m         2023-01-17
-      11   Stephanie Garcia Mayen         Tristan Thomas  40   m         2023-01-16
-      12             Cade Amerson Nohelia Kills In Sight  83   f         2023-01-21
-      13             Cade Amerson       Sage Wachtendorf  90   m         2023-01-21
-      14   Nohelia Kills In Sight        Faalih al-Saleh  35   m         2023-01-28
-      15          Faalih al-Saleh            Nasteh Long  43   f         2023-02-01
-      16          Faalih al-Saleh        Dominica Abeyta   1   f         2023-01-30
-      17          Faalih al-Saleh         Patrick Martin  29   m         2023-02-02
-      18          Dominica Abeyta          Bryant Taylor  78   m         2023-02-03
-      19          Dominica Abeyta        Arkaan al-Faris  22   m         2023-02-04
-      20          Dominica Abeyta          Olivia Carson  70   f         2023-02-03
+                       from                 to age sex date_first_contact
+      1      Sarah Donnelly        Mariah Gray  58   f         2022-12-27
+      2      Sarah Donnelly  Mickenzie Cunniff  29   f         2022-12-30
+      3   Mickenzie Cunniff     Malinda Kelley  24   f         2023-01-12
+      4   Mickenzie Cunniff      Ulises Garcia  42   m         2023-01-11
+      5   Mickenzie Cunniff      Sarah Carlson  48   f         2023-01-09
+      6      Malinda Kelley     Fallon Estrada  76   f         2023-01-11
+      7      Malinda Kelley     Victor Carreon  39   m         2023-01-14
+      8      Malinda Kelley Jesse Breckenridge  24   m         2023-01-12
+      9      Fallon Estrada    Naqaa al-Sarwar  53   f         2023-01-20
+      10 Jesse Breckenridge     Israa al-Bahri  86   f         2023-01-17
+      11 Jesse Breckenridge    Bianca Gallegos  40   f         2023-01-16
+      12     Israa al-Bahri        Jacob Hailu  83   m         2023-01-21
+      13     Israa al-Bahri Amatullah el-Laham  90   f         2023-01-21
+      14        Jacob Hailu     Austin Larabee  35   m         2023-01-28
+      15     Austin Larabee   Reyesa Hernandez  43   f         2023-02-01
+      16     Austin Larabee        Ha-Mi Zheng   1   f         2023-01-30
+      17     Austin Larabee  Siraaj el-Massoud  29   m         2023-02-02
+      18        Ha-Mi Zheng         Cade Jumbo  78   m         2023-02-03
+      19        Ha-Mi Zheng   Kalyn Fredericks  22   f         2023-02-04
[…]

[…]
