+++ tests/testthat/_snaps/sim_linelist.md
@@ -3,165 +3,165 @@
     Code
       sim_linelist()
     Output
-           id                   case_name case_type sex age date_onset date_reporting
-      1     1            Lolette Phillips suspected   f  59 2023-01-01     2023-01-01
-      2     2                  James Jack suspected   m  90 2023-01-01     2023-01-01
-      3     3                 Chen Kantha confirmed   m   4 2023-01-02     2023-01-02
-      4     5             Saleema al-Zaki  probable   f  29 2023-01-04     2023-01-04
-      5     6                David Ponzio confirmed   m  14 2023-01-05     2023-01-05
-      6     7            Christopher Ward  probable   m  85 2023-01-06     2023-01-06
-      7    10                 Laura Ilaoa confirmed   f  25 2023-01-13     2023-01-13
-      8    11                Morgan Mason suspected   f  34 2023-01-11     2023-01-11
-      9    12            Cornelius Turner confirmed   m  89 2023-01-13     2023-01-13
-      10   14          Shaddaad el-Younes suspected   m  63 2023-01-11     2023-01-11
-      11   18         Carlo Ceazar Corpuz confirmed   m  74 2023-01-14     2023-01-14
-      12   19              Bishr al-Qasim confirmed   m  34 2023-01-12     2023-01-12
-      13   22           Nichole Mcfarland confirmed   f  29 2023-01-12     2023-01-12
-      14   24               Jubair el-Aly confirmed   m  46 2023-01-13     2023-01-13
-      15   27            Hafsa al-Massoud  probable   f  24 2023-01-14     2023-01-14
-      16   29                Trever Manis confirmed   m  41 2023-01-21     2023-01-21
-      17   30          Maiya Vathanavarin suspected   f  72 2023-01-19     2023-01-19
-      18   31            Ghaamid el-Zafar  probable   m   6 2023-01-20     2023-01-20
-      19   33      Kelly Runningbear-Paul confirmed   f  61 2023-01-14     2023-01-14
-      20   34                James Welter  probable   m  44 2023-01-14     2023-01-14
-      21   35            Austin Manheimer confirmed   m  71 2023-01-27     2023-01-27
-      22   38         Orville Stadelbauer confirmed   m  69 2023-01-23     2023-01-23
-      23   39                Jennifer Ouk  probable   f  61 2023-01-23     2023-01-23
-      24   41            Nafeesa al-Zaher confirmed   f  64 2023-01-24     2023-01-24
-      25   44           Nawwaara el-Pasha  probable   f  61 2023-01-22     2023-01-22
-      26   45          Saleet el-Mohammed  probable   m  66 2023-01-22     2023-01-22
-      27   47           Christian Mendoza  probable   m  81 2023-01-28     2023-01-28
-      28   50                Eraina Mcgee confirmed   f  53 2023-01-28     2023-01-28
-      29   51                   Mia Floyd  probable   f  45 2023-01-29     2023-01-29
-      30   53               Collin Chavez confirmed   m  42 2023-02-02     2023-02-02
-      31   54              Itzel Gonzales confirmed   f  78 2023-02-03     2023-02-03
-      32   55            Alexis Caramillo confirmed   f  35 2023-01-24     2023-01-24
-      33   56             Ronnie Scurlock suspected   m  54 2023-01-26     2023-01-26
-      34   58                 Shane Ahmad confirmed   m  68 2023-01-25     2023-01-25
-      35   59         Ghaaliya al-Shaheed  probable   f   3 2023-01-23     2023-01-23
-      36   60               Exzinia Smith  probable   f  44 2023-01-29     2023-01-29
[…]
