### 2025-11-10 4b81aea
add info alert about proportion of m/f in vis-linelist vignette
