### 2025-11-11 3ef2c3e
update zenodo DOI in README to version agnostic DOI
