+++ reference/sim_outbreak.md
@@ -1,0 +1,401 @@
+# Simulate a line list and a contacts table
+
+The line list and contacts are simulated using a branching process and
+parameterised with epidemiological parameters.
+
+## Usage
+
+``` r
+sim_outbreak(
+  contact_distribution = function(x) stats::dpois(x = x, lambda = 2),
+  infectious_period = function(x) stats::rlnorm(n = x, meanlog = 2, sdlog = 0.5),
+  prob_infection = 0.5,
+  onset_to_hosp = function(x) stats::rlnorm(n = x, meanlog = 1.5, sdlog = 0.5),
+  onset_to_death = function(x) stats::rlnorm(n = x, meanlog = 2.5, sdlog = 0.5),
+  onset_to_recovery = NULL,
+  reporting_delay = NULL,
+  hosp_risk = 0.2,
+  hosp_death_risk = 0.5,
+  non_hosp_death_risk = 0.05,
+  outbreak_start_date = as.Date("2023-01-01"),
+  anonymise = FALSE,
+  outbreak_size = c(10, 10000),
+  population_age = c(1, 90),
+  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
+  contact_tracing_status_probs = c(under_followup = 0.7, lost_to_followup = 0.2, unknown
+    = 0.1),
+  config = create_config()
+)
+```
+
+## Arguments
+
+- contact_distribution:
+
+  A `function` or an `<epiparameter>` object to generate the number of
+  contacts per infection.
+
+  The function can be defined or anonymous. The function must have a
+  single argument in the form of an `integer` vector with elements
+  representing the number of contacts, and return a `numeric` vector
+  where each element corresponds to the probability of observing the
+  number of contacts in the vector passed to the function. The index of
+  the `numeric` vector returned is offset by one to the corresponding
+  probability of observing the number of contacts, i.e. the first
+  element of the output vector is the probability of observing zero
+  contacts, the second element is the probability of observing one
+  contact, etc.
+
+  An `<epiparameter>` can be provided. This will be converted into a
+  probability mass function internally.
+
+  The default is an anonymous function with a Poisson probability mass
+  function ([`dpois()`](https://rdrr.io/r/stats/Poisson.html)) with a
+  mean (\\\lambda\\) of 2 contacts per infection.
+
+- infectious_period:
+
+  A `function` or an `<epiparameter>` object for the infectious period.
+  This defines the duration from becoming infectious to no longer
+  infectious. In the simulation, individuals are assumed to become
+  infectious immediately after being infected (the latency period is
+  assumed to be zero). The time intervals between an infected individual
+  and their contacts are assumed to be uniformly distributed within the
+  infectious period. Infectious periods must be strictly positive.
+
+  The function can be defined or anonymous. The function must return a
+  vector of randomly generated real numbers representing sampled
+  infectious periods. The function must have a single argument, the
+  number of random infectious periods to generate.
+
+  An `<epiparameter>` can be provided. This will be converted into
+  random number generator internally.
+
+  The default is an anonymous function with a lognormal distribution
+  random number generator
+  ([`rlnorm()`](https://rdrr.io/r/stats/Lognormal.html)) with
+  `meanlog = 2` and `sdlog = 0.5`.
+
+- prob_infection:
+
+  A single `numeric` for the probability of a secondary contact being
+  infected by an infected primary contact.
+
+- onset_to_hosp:
+
+  A `function` or an `<epiparameter>` object for the
+  onset-to-hospitalisation delay distribution. `onset_to_hosp` can also
+  be set to `NULL` to not simulate hospitalisation (admission) dates.
+
[…]
