+++ reference/dot-check_sim_input.md
@@ -1,0 +1,259 @@
+# Check if arguments input to simulation function are valid
+
+Check if arguments input to simulation function are valid
+
+## Usage
+
+``` r
+.check_sim_input(
+  sim_type = c("linelist", "contacts", "outbreak"),
+  contact_distribution,
+  infectious_period,
+  prob_infection,
+  outbreak_start_date,
+  outbreak_size,
+  onset_to_hosp = NULL,
+  onset_to_death = NULL,
+  onset_to_recovery = NULL,
+  anonymise = NULL,
+  case_type_probs = NULL,
+  contact_tracing_status_probs = NULL,
+  hosp_risk = NULL,
+  hosp_death_risk = NULL,
+  non_hosp_death_risk = NULL,
+  population_age = NULL
+)
+```
+
+## Arguments
+
+- sim_type:
+
+  A `character` string specifying which simulation function this
+  function is being called within.
+
+- contact_distribution:
+
+  A `function` or an `<epiparameter>` object to generate the number of
+  contacts per infection.
+
+  The function can be defined or anonymous. The function must have a
+  single argument in the form of an `integer` vector with elements
+  representing the number of contacts, and return a `numeric` vector
+  where each element corresponds to the probability of observing the
+  number of contacts in the vector passed to the function. The index of
+  the `numeric` vector returned is offset by one to the corresponding
+  probability of observing the number of contacts, i.e. the first
+  element of the output vector is the probability of observing zero
+  contacts, the second element is the probability of observing one
+  contact, etc.
+
+  An `<epiparameter>` can be provided. This will be converted into a
+  probability mass function internally.
+
+  The default is an anonymous function with a Poisson probability mass
+  function ([`dpois()`](https://rdrr.io/r/stats/Poisson.html)) with a
+  mean (\\\lambda\\) of 2 contacts per infection.
+
+- infectious_period:
+
+  A `function` or an `<epiparameter>` object for the infectious period.
+  This defines the duration from becoming infectious to no longer
+  infectious. In the simulation, individuals are assumed to become
+  infectious immediately after being infected (the latency period is
+  assumed to be zero). The time intervals between an infected individual
+  and their contacts are assumed to be uniformly distributed within the
+  infectious period. Infectious periods must be strictly positive.
+
+  The function can be defined or anonymous. The function must return a
+  vector of randomly generated real numbers representing sampled
+  infectious periods. The function must have a single argument, the
+  number of random infectious periods to generate.
+
+  An `<epiparameter>` can be provided. This will be converted into
+  random number generator internally.
+
+  The default is an anonymous function with a lognormal distribution
+  random number generator
+  ([`rlnorm()`](https://rdrr.io/r/stats/Lognormal.html)) with
+  `meanlog = 2` and `sdlog = 0.5`.
+
+- prob_infection:
+
+  A single `numeric` for the probability of a secondary contact being
+  infected by an infected primary contact.
+
+- outbreak_start_date:
+
+  A `date` for the start of the outbreak.
+
+- outbreak_size:
+
+  A `numeric` vector of length 2 defining the minimum and the maximum
+  number of infected individuals for the simulated outbreak. Default is
+  `c(10, 1e4)`, so the minimum outbreak size is 10 infected individuals,
+  and the maximum outbreak size is 10,000 infected individuals. Either
+  number can be changed to increase or decrease the maximum or minimum
+  outbreak size to allow simulating larger or smaller outbreaks. If the
+  minimum outbreak size cannot be reached after running the simulation
+  for many iterations (internally) then the function errors, whereas if
+  the maximum outbreak size is exceeded the function returns the data
[…]
