+++ authors.md
@@ -1,0 +1,46 @@
+# Authors and Citation
+
+## Authors
+
+- **Joshua W. Lambert**. Author, maintainer, copyright holder.
+  [](https://orcid.org/0000-0001-5218-3046)
+
+- **Carmen Tamayo Cuartero**. Author.
+  [](https://orcid.org/0000-0003-4184-2864)
+
+- **Hugo Gruson**. Contributor, reviewer.
+  [](https://orcid.org/0000-0002-4094-1476)
+
+- **Pratik R. Gupte**. Contributor, reviewer.
+  [](https://orcid.org/0000-0001-5294-7819)
+
+- **Adam Kucharski**. Reviewer.
+  [](https://orcid.org/0000-0001-8814-9421)
+
+- **Chris Hartgerink**. Reviewer.
+  [](https://orcid.org/0000-0003-1050-6809)
+
+- **Sebastian Funk**. Contributor.
+  [](https://orcid.org/0000-0002-2842-3406)
+
+- **London School of Hygiene and Tropical Medicine, LSHTM**. Copyright
+  holder.
+  [![ROR](https://raw.githubusercontent.com/ror-community/ror-logos/main/ror-icon-rgb.svg)](https://ror.org/00a0jsq62)
+
+## Citation
+
+Source:
+[`inst/CITATION`](https://github.com/epiverse-trace/simulist/blob/main/inst/CITATION)
+
+Lambert J, Tamayo Cuartero C (2025). *simulist: Simulate Disease
+Outbreak Line List and Contacts Data*.
+[doi:10.5281/zenodo.10471458](https://doi.org/10.5281/zenodo.10471458),
+<https://epiverse-trace.github.io/simulist/>.
+
+    @Manual{,
+      title = {simulist: Simulate Disease Outbreak Line List and Contacts Data},
+      author = {Joshua W. Lambert and Carmen {Tamayo Cuartero}},
+      year = {2025},
+      doi = {10.5281/zenodo.10471458},
+      url = {https://epiverse-trace.github.io/simulist/},
+    }

[…]
