### 2025-11-12 809e64d
use n instead of x for function signature of random sampling functions
