### 2025-11-12 2d56b8f
use n instead of x for function signature of random sampling functions
