+++ vignettes/reporting-delays-truncation.Rmd
@@ -106,7 +106,7 @@ linelist <- sim_linelist(
   prob_infection = 0.5,
   onset_to_hosp = onset_to_hosp,
   onset_to_death = onset_to_death,
-  reporting_delay = function(x) rlnorm(n = x, meanlog = 1, sdlog = 1)
+  reporting_delay = function(n) rlnorm(n = n, meanlog = 1, sdlog = 1)
 )
 head(linelist)
 ```
@@ -198,7 +198,7 @@ linelist <- sim_linelist(
   prob_infection = 0.5,
   onset_to_hosp = onset_to_hosp,
   onset_to_death = onset_to_death,
-  reporting_delay = function(x) rep(5, times = x)
+  reporting_delay = function(n) rep(5, times = n)
 )
 head(linelist)
 linelist$date_reporting - linelist$date_onset
@@ -223,7 +223,7 @@ linelist <- sim_linelist(
   prob_infection = 0.5,
   onset_to_hosp = onset_to_hosp,
   onset_to_death = onset_to_death,
-  reporting_delay = function(x) rlnorm(n = x, meanlog = 2, sdlog = 0.5)
+  reporting_delay = function(n) rlnorm(n = n, meanlog = 2, sdlog = 0.5)
 )
 
 # first 6 rows of linelist
@@ -410,7 +410,7 @@ linelist <- sim_linelist(
   prob_infection = 0.5,
   onset_to_hosp = onset_to_hosp,
   onset_to_death = onset_to_death,
-  reporting_delay = function(x) rlnorm(n = x, meanlog = 2, sdlog = 0.5),
+  reporting_delay = function(n) rlnorm(n = n, meanlog = 2, sdlog = 0.5),
   outbreak_size = c(500, 5000)
 )
 ```
