### 2025-11-19 9605fc1
remove .add_reporting_delay function
