### 2025-11-19 7397130
call reporting_delay in .sim_internal
