### 2025-11-19 c27a16b
update as_function to accept and convert reporting_delay
