### 2025-11-19 ee9f8e7
update .check_sim_input to check reporting_delay
