+++ R/checkers.R
@@ -106,6 +106,7 @@
                              onset_to_hosp = NULL,
                              onset_to_death = NULL,
                              onset_to_recovery = NULL,
+                             reporting_delay = NULL,
                              anonymise = NULL,
                              case_type_probs = NULL,
                              contact_tracing_status_probs = NULL,
@@ -131,6 +132,18 @@
     .check_func_req_args(onset_to_hosp, func_name = "onset_to_hosp")
     .check_func_req_args(onset_to_death, func_name = "onset_to_death")
     .check_func_req_args(onset_to_recovery, func_name = "onset_to_recovery")
+    .check_func_req_args(reporting_delay, func_name = "reporting_delay")
+
+    # check if reporting_delay generates a number
+    reporting_delay_eval <- reporting_delay(1)
+    if (!is.numeric(reporting_delay_eval)) {
+      stop(
+        "The `reporting_delay` must be NULL or a function that ",
+        "generates numbers.",
+        call. = FALSE
+      )
+    }
+
     checkmate::assert_logical(anonymise, len = 1)
     checkmate::assert_numeric(case_type_probs, len = 3, lower = 0, upper = 1)
     checkmate::assert_names(
