### 2025-11-19 4ef0349
update sim_linelist and sim_outbreak to pass reporting_delay to as_function and .check_sim_input
