+++ R/sim_linelist.R
@@ -341,7 +341,8 @@ sim_linelist <- function(contact_distribution = function(x) stats::dpois(x = x, 
       infectious_period = infectious_period,
       onset_to_hosp = onset_to_hosp,
       onset_to_death = onset_to_death,
-      onset_to_recovery = onset_to_recovery
+      onset_to_recovery = onset_to_recovery,
+      reporting_delay = reporting_delay
     )
   )
   contact_distribution <- funcs$contact_distribution
@@ -349,6 +350,7 @@ sim_linelist <- function(contact_distribution = function(x) stats::dpois(x = x, 
   onset_to_hosp <- funcs$onset_to_hosp
   onset_to_death <- funcs$onset_to_death
   onset_to_recovery <- funcs$onset_to_recovery
+  reporting_delay <- funcs$reporting_delay
 
   .check_sim_input(
     sim_type = "linelist",
@@ -360,6 +362,7 @@ sim_linelist <- function(contact_distribution = function(x) stats::dpois(x = x, 
     onset_to_hosp = onset_to_hosp,
     onset_to_death = onset_to_death,
     onset_to_recovery = onset_to_recovery,
+    reporting_delay = reporting_delay,
     anonymise = anonymise,
     case_type_probs = case_type_probs,
     hosp_risk = hosp_risk,
