### 2025-11-19 76ef914
update add_cols and .check_sim_input function documentation
