### 2025-11-19 02225cb
add reporting_delay to .check_sim_input tests
