### 2025-11-19 ad650cd
add unit test for .check_sim_input error for reporting_delay
