### 2025-11-20 fff34cc
add nolint cyclocomp_linter flag to as_function
