### 2025-11-20 582450b
update nolint flag syntax to specify exclusion
