### 2025-11-20 c58df08
update reporting_delay argument documentation
