### 2025-11-20 dbe4387
update reporting_delay argument documentation
