+++ articles/reporting-delays-truncation.md
@@ -151,27 +151,20 @@ linelist <- sim_linelist(
   reporting_delay = function(n) rlnorm(n = n, meanlog = 1, sdlog = 1)
 )
 head(linelist)
-#>   id            case_name case_type sex age date_onset date_reporting
-#> 1  1       Yoonus al-Azam suspected   m  51 2023-01-01     2023-01-07
-#> 2  2       Kristina Clark confirmed   f  53 2023-01-04     2023-01-06
-#> 3  3         Karen Meehan confirmed   f  16 2023-01-05     2023-01-20
-#> 4  4 Ritika Brandenburger suspected   f  41 2023-01-05     2023-01-09
-#> 5  5       Steven Kaliski confirmed   m   4 2023-01-04     2023-01-09
-#> 6  6          Keely Atene suspected   f  68 2023-01-05     2023-01-12
-#>   date_admission   outcome date_outcome date_first_contact date_last_contact
-#> 1     2023-01-03      died   2023-01-26               <NA>              <NA>
-#> 2           <NA> recovered         <NA>         2022-12-28        2023-01-03
-#> 3           <NA> recovered         <NA>         2022-12-29        2023-01-06
-#> 4           <NA> recovered         <NA>         2022-12-31        2023-01-05
-#> 5           <NA> recovered         <NA>         2023-01-02        2023-01-05
-#> 6           <NA> recovered         <NA>         2023-01-02        2023-01-06
-#>   ct_value
-#> 1       NA
-#> 2     26.1
-#> 3     26.6
-#> 4       NA
-#> 5     25.0
-#> 6       NA
+#>   id       case_name case_type sex age date_onset date_reporting date_admission
+#> 1  1   Tiarra Morris confirmed   f  42 2023-01-01     2023-01-02           <NA>
+#> 2  4     Keiara Yang suspected   f  63 2023-01-04     2023-01-08           <NA>
+#> 3  7  Monika Watters confirmed   f  83 2023-01-01     2023-01-02           <NA>
+#> 4 11      Tina Zhang confirmed   f  21 2023-01-03     2023-01-09           <NA>
+#> 5 13 Nicholas Polaco  probable   m  36 2023-01-05     2023-01-11           <NA>
+#> 6 14    Vidya Nguyen confirmed   f  84 2023-01-06     2023-01-08           <NA>
+#>     outcome date_outcome date_first_contact date_last_contact ct_value
+#> 1 recovered         <NA>               <NA>              <NA>     24.1
+#> 2 recovered         <NA>         2022-12-25        2023-01-06       NA
+#> 3 recovered         <NA>         2022-12-30        2023-01-04     23.9
+#> 4 recovered         <NA>         2022-12-30        2023-01-03     29.7
+#> 5 recovered         <NA>         2023-01-01        2023-01-07       NA
+#> 6 recovered         <NA>         2022-12-31        2023-01-09     24.0
 ```
 
 Here from the first 6 rows of the line list you can see differences
@@ -214,9 +207,9 @@ ggplot(data = tidy_linelist) +
   ) +
   theme_bw() +
   theme(legend.position = "bottom", axis.text.y = element_text(size = 4))
[…]
