### 2025-11-20 b429052
fix nolint flag linter exclusion syntax
