+++ vignettes/reporting-delays-truncation.Rmd
@@ -128,8 +128,10 @@ tidy_linelist <- linelist %>%
   pivot_longer(
     cols = c("date_onset", "date_reporting", "date_admission", "date_outcome")
   ) %>%
-  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>% # nolint consecutive_mutate_linter
-  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)) # nolint consecutive_mutate_linter
+  # nolint start: consecutive_mutate_linter.
+  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>%
+  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE))
+  # nolint end
 
 ggplot(data = tidy_linelist) +
   geom_line(
@@ -242,8 +244,10 @@ tidy_linelist <- linelist %>%
   pivot_longer(
     cols = c("date_onset", "date_reporting", "date_admission", "date_outcome")
   ) %>%
-  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>% # nolint consecutive_mutate_linter
-  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)) # nolint consecutive_mutate_linter
+  # nolint start: consecutive_mutate_linter.
+  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>%
+  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE))
+  # nolint end
 
 truncation_day <- 14
 trunc_date <- max(tidy_linelist$value, na.rm = TRUE) - truncation_day
