### 2025-11-20 7b5fffa
fix nolint flag linter exclusion syntax
