+++ vignettes/wrangling-linelist.Rmd
@@ -269,7 +269,7 @@ Not every column in the simulated line list may be required for the use case at 
 
 ```{r, rm-ct-col-tidyverse}
 # remove column by name
-linelist %>% # nolint one_call_pipe_linter
+linelist %>% # nolint: one_call_pipe_linter.
   select(!ct_value)
 ```
 
