### 2025-11-20 b9e7efc
merge mutate calls in reporting-delays-truncation.Rmd vignette
