+++ vignettes/reporting-delays-truncation.Rmd
@@ -128,10 +128,10 @@ tidy_linelist <- linelist %>%
   pivot_longer(
     cols = c("date_onset", "date_reporting", "date_admission", "date_outcome")
   ) %>%
-  # nolint start: consecutive_mutate_linter.
-  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>%
-  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE))
-  # nolint end
+  mutate(
+    ordering_value = ifelse(name == "date_onset", value, NA),
+    case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)
+  )
 
 ggplot(data = tidy_linelist) +
   geom_line(
@@ -244,10 +244,10 @@ tidy_linelist <- linelist %>%
   pivot_longer(
     cols = c("date_onset", "date_reporting", "date_admission", "date_outcome")
   ) %>%
-  # nolint start: consecutive_mutate_linter.
-  mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>%
-  mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE))
-  # nolint end
+  mutate(
+    ordering_value = ifelse(name == "date_onset", value, NA),
+    case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)
+  )
 
 truncation_day <- 14
 trunc_date <- max(tidy_linelist$value, na.rm = TRUE) - truncation_day
