### 2025-11-20 2d338f2
use echo in Rmarkdown chunks to hide nolint flags
