### 2025-11-20 3df88e9
use echo in Rmarkdown chunks to hide nolint flags
