### 2025-11-20 440efc0
increment to development version 0.6.0.9000
