### 2025-11-20 d7bf452
Update `CITATION.cff`
