+++ dev/articles/index.md
@@ -1,0 +1,24 @@
+# Articles
+
+### Customising simulated outbreak
+
+- [Age-stratified hospitalisation and death
+  risks](https://epiverse-trace.github.io/simulist/dev/articles/age-strat-risks.md):
+- [Age-structured
+  population](https://epiverse-trace.github.io/simulist/dev/articles/age-struct-pop.md):
+- [Time-varying case fatality
+  risk](https://epiverse-trace.github.io/simulist/dev/articles/time-varying-cfr.md):
+- [Reporting delays and right-truncation of line list
+  data](https://epiverse-trace.github.io/simulist/dev/articles/reporting-delays-truncation.md):
+
+### Wrangling and plotting data
+
+- [Wrangling simulated outbreak
+  data](https://epiverse-trace.github.io/simulist/dev/articles/wrangling-linelist.md):
+- [Visualising simulated
+  data](https://epiverse-trace.github.io/simulist/dev/articles/vis-linelist.md):
+
+### Developer Documentation
+
+- [Design Principles for
+  {simulist}](https://epiverse-trace.github.io/simulist/dev/articles/design-principles.md):
