+++ dev/articles/reporting-delays-truncation.md
@@ -1,0 +1,692 @@
+# Reporting delays and right-truncation of line list data
+
+If you are unfamiliar with the {simulist} package or the
+[`sim_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_linelist.md)
+function [Get Started
+vignette](https://epiverse-trace.github.io/simulist/dev/articles/simulist.md)
+is a great place to start.
+
+This vignette covers reporting delays in line list data, how to use
+{simulist} to produce line list data with a delay between symptom onset
+and reporting, explains right truncation in outbreak data sets and shows
+how to use the
+[`truncate_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/truncate_linelist.md)
+function in {simulist} to augment a simulated line list to resemble
+real-time outbreak data.
+
+## Reporting delays
+
+Reporting delays in line list data during an infectious disease outbreak
+refer to the time lag between when an event (e.g., infection, symptom
+onset, or hospitalisation) occurs and when it is officially recorded or
+reported. These delays can arise from factors like diagnostic testing
+turnaround times, administrative processing, or delayed
+healthcare-seeking behaviour. In line list data sets, these delays are
+captured as separate times/dates for different event types, such as the
+dates of symptom onset, specimen collection, laboratory confirmation, or
+case reporting. For each case, the delay is the difference between these
+times/dates, allowing researchers to quantify and adjust for reporting
+delays when analysing outbreak dynamics. Accurate recording of these
+dates is crucial for reconstructing the epidemic curve and assessing the
+timeliness of public health responses.
+
+In {simulist},
+[`sim_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_linelist.md)
+outputs a line list with a date of reporting (`$date_reporting` column),
+which by default will be identical to the date of symptom onset
+(`$date_onset`). However, it is possible to parameterise a reporting
+delay that can be used by
+[`sim_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_linelist.md)
+to simulate a delay between symptom onset and reporting. In {simulist}
+we assume that the reporting delay is the delay between symptom onset
+and it being reported, it does not reflect reporting of hospitalisation
+or outcome times.
+
+``` r
+library(simulist)
+library(epiparameter)
+library(tidyr)
+library(dplyr)
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
+library(ggplot2)
+library(incidence2)
+#> Loading required package: grates
+```
+
+First we load the required delay distributions using the {epiparameter}
+package.
+
+``` r
+contact_distribution <- epiparameter(
+  disease = "COVID-19",
+  epi_name = "contact distribution",
+  prob_distribution = create_prob_distribution(
+    prob_distribution = "pois",
+    prob_distribution_params = c(mean = 2)
+  )
+)
+#> Citation cannot be created as author, year, journal or title is missing
+
+infectious_period <- epiparameter(
+  disease = "COVID-19",
+  epi_name = "infectious period",
+  prob_distribution = create_prob_distribution(
+    prob_distribution = "gamma",
+    prob_distribution_params = c(shape = 3, scale = 2)
+  )
+)
[…]
