### 2025-11-20 0ac6454
add event timeline example to vis-linelist.Rmd vignette
