### 2025-11-20 163a2ce
add event timeline example to vis-linelist.Rmd vignette
