### 2025-11-20 4b8cbc5
update JOSS paper from reviewer 1 comments
