+++ joss/paper.md
@@ -31,51 +31,51 @@ bibliography: paper.bib
 
 ![simulist package hex logo, containing the words simulist and underneath, powered by data.org, with overlapping rectangles above the words](../man/figures/logo.svg){ width=240px }\
 
-Epidemic and pandemic preparedness and analytics response requires robust analysis methods and a deep understanding of outbreak data. We introduce `{simulist}`, an open-source R package for simulating realistic infectious disease outbreak data. It is designed to allow users to generate data with specific disease outbreak characteristics by enabling flexible parameterisation of processes and variables, in particular, infectious disease epidemiology (e.g. delay distributions), contact patterns and demographic information, to simulate two common datasets in outbreak settings: 1) line list data, and 2) contact tracing data. It also offers post-processing of line list data to replicate right-truncation of real-time outbreak data, as well as creating "messy" data with realistically incomplete and inconsistent values. 
+Epidemic and pandemic preparedness and analytics response requires robust analysis methods and a deep understanding of outbreak data. We introduce `simulist`, an open-source R package for simulating realistic infectious disease outbreak data. It is designed to allow users to generate data with specific disease outbreak characteristics by enabling flexible parameterisation of processes and variables, in particular, infectious disease epidemiology (e.g. delay distributions), contact patterns and demographic information, to simulate two common datasets in outbreak settings: 1) line list data, and 2) contact tracing data. It also offers post-processing of line list data to replicate right-truncation of real-time outbreak data, as well as creating "messy" data with realistically incomplete and inconsistent values. 
 
 ## Statement of need
 
 Synthetic data, which is generated by a known simulation process, is useful in many scenarios where a "ground truth" is required, including teaching concepts on data collection, data cleaning and data visualisation; and testing analysis methods for precision, adequacy and robustness. In the analysis of infectious disease outbreaks, termed _outbreak analytics_ [@polonskyOutbreakAnalyticsDeveloping2019a; @jombartWhyDevelopmentOutbreak2021; @carterHowImproveOutbreak2021], one of the most common forms of data is a line list. This is a tabular dataset where each row corresponds to a single case in the outbreak, recording information on person, place and time. Information for each case can include, but is not limited to, personal information (name, age, sex), key event dates (date of symptom onset, date of hospitalisation, date of outcome), clinical outcomes, and geographic information (coordinates, administrative area). Understanding and characterising the transmissibility and severity of infectious disease outbreaks typically use line list data [@coriKeyDataOutbreak2017]. 
 
-In the ecosystem of R packages for epidemiology, there are a range of tools for simulating and analysing line list data [@jombartCRANTaskView2025], however, there are no stable, actively maintained and tested R packages that simulate epidemiologically realistic line list data[^1]. The `{simulist}` R package offers functionality to simulate outbreak data (line list and contact tracing) with options for users to define the epidemiological and demographic characteristics of the outbreak, while also being designed on the principle of maximum interoperability with other related R packages, e.g. `{epicontacts}` [@campbellEpicontactsHandlingVisualisation2017], `{epiparameter}` [@lambertEpiparameterClassesHelper2025], and `{cleanepi}` [@maneCleanepiCleanStandardize2024]. Accepting epidemiological parameters directly from `{epiparameter}`, an R package with a library of parameter estimates from the literature, facilitates generating synthetic data with the characteristics of past outbreaks.
[…]
