### 2025-11-20 3383e72
update JOSS paper.bib
