+++ joss/paper.bib
@@ -138,6 +138,22 @@
   urldate = {2025-03-07}
 }
 
+@article{gallagherIdentificationAttributionWeekly2025,
+  title = {Identification and Attribution of Weekly Periodic Biases in Global Epidemiological Time Series Data},
+  author = {Gallagher, Kit and Creswell, Richard and Gavaghan, David and Lambert, Ben},
+  year = {2025},
+  month = feb,
+  journal = {BMC Research Notes},
+  volume = {18},
+  number = {1},
+  pages = {78},
+  issn = {1756-0500},
+  doi = {10.1186/s13104-025-07145-y},
+  urldate = {2025-11-20},
+  abstract = {Abstract                            Objective               COVID-19 data exhibit various biases, not least a significant weekly periodic oscillation observed in case and death data from multiple countries. There has been debate over whether this may be attributed to weekly socialising and working patterns, or is due to underlying biases in the reporting process. We investigate these periodic reporting trends in epidemics of COVID-19 and cholera, and discuss the possible origin of these oscillations.                                         Results               We present a systematic, global characterisation of these weekly biases and identify an equivalent bias in the current Haitian cholera outbreak. By comparing published COVID-19 time series to retrospective datasets from the United Kingdom (UK), we demonstrate that the weekly trends observed in the UK may be fully explained by biases in the testing and reporting processes. These conclusions play an important role in forecasting healthcare demand and determining suitable interventions for future infectious disease outbreaks.},
+  langid = {english}
+}
+
 @misc{jombartCRANTaskView2025,
   title = {{{CRAN Task View}}: {{Epidemiology}}},
   author = {Jombart, Thibaut and Rolland, Matthieu and Gruson, Hugo},
