### 2025-11-20 a907fdf
remove explicit mention of *_risk argument from JOSS paper
