### 2025-11-20 ef66bf6
try and fix \ref
