### 2025-11-20 1e501e8
try \autoref with a single image
