+++ joss/paper.md
@@ -49,7 +49,7 @@ However, to our knowledge, there are no existing tools that offer an easy method
 
 `simulist` offers two modules of functionality: 1) simulation and 2) post-processing. In the simulation module there is a function for simulating a line list, a function for simulating contact tracing data, and a function for simulating both. Respectively these output, an outbreak line list (`data.frame`), a contact tracing data set (`data.frame`), and lastly both an outbreak line list and contact tracing data set that are linked (i.e. from the same outbreak) (`list` of `data.frame`s). All simulation functions internally call a branching process model [@farringtonBranchingProcessModels2003a] with the option of network adjusted contact patterns to account for the fact that choosing someone at random by following up a contact chooses individuals with a probability proportional to their number of contacts [@alexanderRiskFactorsEvolutionary2010; @blumbergDetectingDifferentialTransmissibilities2014]. No data is stored in the package, all outbreak data is generated upon calling a simulation function. All simulation functions can be run without requiring the user to specify any parameters (i.e. function arguments, see function documentation for defaults), but all functions offer a range of arguments to modify the characteristics of the outbreak data. The parameters that can be specified by the user are organised into two levels. The parameters that control: the branching process, key epidemiological delays, risks, and demographic parameters are function arguments because they are more likely to be adjusted by users. Parameters thought to rarely require changing, except for advanced use cases are passed to the function via a configuration list, to prevent the function signature being overly complex. A helper function is provided to automatically generate the configuration list with sensible defaults, and allow easy modification, and this is called by default in the simulation functions.
 
-![Part a. The daily incidence of cases, hospitalisations and deaths, facetted by sex for a simulated outbreak.](incidence.png){ width=33% } ![Part b. A contact network for a simulated outbreak](network.png){ width=33% } ![Part c. A multi-dot plot for the events in a simulated line list. \label{fig:simulist}](multi-dot-plot.png){ width=33% }
+![The daily incidence of cases, hospitalisations and deaths, facetted by sex for a simulated outbreak.\label{fig:simulist}](incidence.png)
 
 Figure 1: (a) daily incidence of symptom onset, hospital admission and deaths facetted by sex for a simulated line list and aggregated and plotted with `incidence2` [@taylorIncidence2ComputeHandle2020]. (b) an `<epicontacts>` network plotted after converted from a simulated line list and contact tracing dataset (this network is interactive when rendered in an IDE or online). (c) a connected dot plot of the events in a simulated line list. All code to reproduce the plots can be found at https://github.com/epiverse-trace/simulist/blob/joss/joss/joss-figs.R, and more examples of these types of plots can be found in the [Visualising simulated data vignette](https://epiverse-trace.github.io/simulist/articles/vis-linelist.html).
 
@@ -67,7 +67,7 @@ Other functionality is documented in the [Get Started vignette](https://epiverse
 
 ## Package Design
 
-The `simulist` package is developed around certain design principles: _interoperability_ with other R packages for upstream (e.g. `epiparameter`) or downstream (e.g. `epicontacts` Figure~\ref{fig:simulist}) use; _ease of use_ of functions (all simulation functions have sensible defaults for use out-of-the-box); and _transparency_, all design decisions are documented in the [Design Principles vignette](https://epiverse-trace.github.io/simulist/articles/design-principles.html) and help developer onboarding. Simulation functions output a `data.frame` (or `list` of `data.frame`s) enabling users to easily transform and manipulate the data, for example with the tidyverse R packages; and post-processing functions are compatible with `data.frame` subclasses (e.g. `<tibble>` or `<data.table>`). 
[…]
