### 2025-11-20 5799ce5
remove figure legend
