### 2025-11-20 6dbd6de
try autoref with smaller fig width
