### 2025-11-20 6d2a1fd
split into three plots
