### 2026-01-01 a08fb1e
docs(license): update copyright year(s)
