+++ dev/reference/create_config.md
@@ -98,20 +98,20 @@ create_config()
 #> $last_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x560b92df67c8>
-#> <environment: 0x560b953e93a0>
+#> <bytecode: 0x55daafe37168>
+#> <environment: 0x55dab1155630>
 #> 
 #> $first_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x560b92df67c8>
-#> <environment: 0x560b953e93a0>
+#> <bytecode: 0x55daafe37168>
+#> <environment: 0x55dab1155630>
 #> 
 #> $ct_distribution
 #> function (n) 
 #> stats::rnorm(n = n, mean = 25, sd = 2)
-#> <bytecode: 0x560b92df6560>
-#> <environment: 0x560b953e93a0>
+#> <bytecode: 0x55daafe36f00>
+#> <environment: 0x55dab1155630>
 #> 
 #> $network
 #> [1] "adjusted"
@@ -127,19 +127,19 @@ create_config(
 #> $last_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x560b92df67c8>
-#> <environment: 0x560b9500f5b8>
+#> <bytecode: 0x55daafe37168>
+#> <environment: 0x55dab0a5e990>
 #> 
 #> $first_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x560b92df67c8>
-#> <environment: 0x560b9500f5b8>
+#> <bytecode: 0x55daafe37168>
+#> <environment: 0x55dab0a5e990>
 #> 
 #> $ct_distribution
 #> function (n) 
 #> rlnorm(n = n, meanlog = 2, sdlog = 1)
-#> <environment: 0x560b93e38c98>
+#> <environment: 0x55dab11f2d58>
 #> 
 #> $network
 #> [1] "adjusted"
