+++ dev/reference/create_config.md
@@ -98,20 +98,20 @@ create_config()
 #> $last_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x55daafe37168>
-#> <environment: 0x55dab1155630>
+#> <bytecode: 0x563e99d287a0>
+#> <environment: 0x563e9d036be8>
 #> 
 #> $first_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x55daafe37168>
-#> <environment: 0x55dab1155630>
+#> <bytecode: 0x563e99d287a0>
+#> <environment: 0x563e9d036be8>
 #> 
 #> $ct_distribution
 #> function (n) 
 #> stats::rnorm(n = n, mean = 25, sd = 2)
-#> <bytecode: 0x55daafe36f00>
-#> <environment: 0x55dab1155630>
+#> <bytecode: 0x563e99d28a40>
+#> <environment: 0x563e9d036be8>
 #> 
 #> $network
 #> [1] "adjusted"
@@ -127,19 +127,19 @@ create_config(
 #> $last_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x55daafe37168>
-#> <environment: 0x55dab0a5e990>
+#> <bytecode: 0x563e99d287a0>
+#> <environment: 0x563e9cd26818>
 #> 
 #> $first_contact_distribution
 #> function (n) 
 #> stats::rpois(n = n, lambda = 3)
-#> <bytecode: 0x55daafe37168>
-#> <environment: 0x55dab0a5e990>
+#> <bytecode: 0x563e99d287a0>
+#> <environment: 0x563e9cd26818>
 #> 
 #> $ct_distribution
 #> function (n) 
 #> rlnorm(n = n, meanlog = 2, sdlog = 1)
-#> <environment: 0x55dab11f2d58>
+#> <environment: 0x563e9d0af578>
 #> 
 #> $network
 #> [1] "adjusted"
