### 2026-01-05 5ac7ab9
add add-paths to update-citation-cff.yaml
