### 2026-01-05 a12eac1
update simulist-package.Rd
