+++ man/simulist-package.Rd
@@ -32,7 +32,7 @@ Other contributors:
   \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [reviewer]
   \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [reviewer]
   \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [contributor]
-  \item London School of Hygiene and Tropical Medicine, LSHTM (00a0jsq62) [copyright holder]
+  \item London School of Hygiene and Tropical Medicine, LSHTM (\href{https://ror.org/00a0jsq62}{ROR}) [copyright holder]
 }
 
 }
