### 2026-01-05 33a4020
Update `CITATION.cff`
