+++ CITATION.cff
@@ -38,7 +38,7 @@ preferred-citation:
     given-names: Carmen
     email: carmen.tamayo-cuartero@lshtm.ac.uk
     orcid: https://orcid.org/0000-0003-4184-2864
-  year: '2025'
+  year: '2026'
   doi: 10.5281/zenodo.10471458
   url: https://epiverse-trace.github.io/simulist/
 repository: https://CRAN.R-project.org/package=simulist
@@ -66,7 +66,7 @@ references:
   institution:
     name: R Foundation for Statistical Computing
     address: Vienna, Austria
-  year: '2025'
+  year: '2026'
   version: '>= 4.2.0'
 - type: software
   title: checkmate
@@ -79,7 +79,7 @@ references:
     given-names: Michel
     email: michellang@gmail.com
     orcid: https://orcid.org/0000-0001-9754-0393
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.checkmate
 - type: software
   title: english
@@ -95,7 +95,7 @@ references:
     given-names: Anthony
   - family-names: Salverda
     given-names: Anne Pier
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.english
 - type: software
   title: epiparameter
@@ -117,7 +117,7 @@ references:
     given-names: Carmen
     email: carmen.tamayo-cuartero@lshtm.ac.uk
     orcid: https://orcid.org/0000-0003-4184-2864
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.epiparameter
   version: '>= 0.4.0'
 - type: software
@@ -131,7 +131,7 @@ references:
     given-names: Tim
     email: tim.taylor@hiddenelephants.co.uk
     orcid: https://orcid.org/0000-0002-8587-7113
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.grates
 - type: software
   title: randomNames
@@ -143,7 +143,7 @@ references:
   - family-names: Betebenner
     given-names: Damian W.
     email: dbetebenner@nciea.org
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.randomNames
 - type: software
   title: rlang
@@ -158,7 +158,7 @@ references:
   - family-names: Wickham
     given-names: Hadley
     email: hadley@posit.co
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.rlang
 - type: software
   title: stats
@@ -169,7 +169,7 @@ references:
   institution:
     name: R Foundation for Statistical Computing
     address: Vienna, Austria
-  year: '2025'
+  year: '2026'
 - type: software
   title: dplyr
   abstract: 'dplyr: A Grammar of Data Manipulation'
@@ -193,7 +193,7 @@ references:
     given-names: Davis
     email: davis@posit.co
     orcid: https://orcid.org/0000-0003-4777-038X
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.dplyr
 - type: software
   title: epicontacts
@@ -224,7 +224,7 @@ references:
   - family-names: Kamvar
     given-names: Zhian N.
     email: zkamvar@gmail.com
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.epicontacts
   version: '>= 1.1.3'
 - type: software
@@ -265,7 +265,7 @@ references:
     given-names: Teun
     name-particle: van den
     orcid: https://orcid.org/0000-0002-9335-7468
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.ggplot2
 - type: software
   title: incidence2
[…]
