### 2026-01-12 #8781 @joshwlambert
Thank you to both reviewer to your comments and to the handling editor for the reminder. I'll open issues and PRs to respond to comments and within this thread with changes. I'll make the changes this week and confirm once all the comments are addressed.
