+++ joss/paper.md
@@ -41,7 +41,7 @@ In the ecosystem of R packages for epidemiology, there are a range of tools for 
 
 ## State of the field
 
-The majority of outbreak analytics uses R [@rcoreteamLanguageEnvironmentStatistical2025] and R packages to answer common epidemiological questions in an outbreak [@tamayocuartero100DayMission2025]. From basic data wrangling and data science using tools such as the tidyverse [@wickhamWelcomeTidyverse2019], to using advanced methods, for example nowcasting/forecasting of transmissibility [@abbottEstimatingTimevaryingReproduction2020a]  This is accompanied by several initiatives that teach epidemiology using R tooling [@batraEpidemiologistHandbook2021; @vallecamposEpiverseTRACETutorials2025] resulting in an active ecosystem of open-source tooling, teaching, and community engagement [@morganWHOHubPandemic2022], for example the [epinowcast community](https://www.epinowcast.org/).
+The majority of outbreak analytics uses R [@rcoreteamLanguageEnvironmentStatistical2025]. There are many R packages that help to answer common epidemiological questions in an outbreak [@tamayocuartero100DayMission2025]. From basic data wrangling and data science tools such as the tidyverse [@wickhamWelcomeTidyverse2019], to using advanced methods, for example nowcasting/forecasting of transmissibility [@abbottEstimatingTimevaryingReproduction2020a]  Additionally, there are several initiatives that teach epidemiology using R tooling [@batraEpidemiologistHandbook2021; @vallecamposEpiverseTRACETutorials2025], resulting in an active ecosystem of open-source tooling, teaching, and community engagement [@morganWHOHubPandemic2022], for example the [epinowcast community](https://www.epinowcast.org/).
 
 However, to our knowledge, there are no existing tools that offer an easy method to generate synthetic outbreak data, with a defined and highly customisable simulation model. The `simulist` R package, part of the Epiverse-TRACE ecosystem [@data.orgEpiverse2022], aims to fill this niche by offering line list simulation functionality that includes: custom parameterisation of epidemiological delay distribution, population-wide or age-stratified hospitalisation and death risks, uniform or age-structured populations, constant or time-varying case fatality risks, and customisable proportions of case types and contact tracing follow-up. This is complimented by post-processing to enable emulation of real-time outbreak dynamics (e.g. right-truncation and outbreak snapshotting) and messying data to mimic the types of issues encounter by people working with empirical outbreak data.
 
@@ -49,11 +49,11 @@ However, to our knowledge, there are no existing tools that offer an easy method
 
[…]
