### 2026-01-16 6f641ad
fix links in figure legends
