### 2026-01-16 a68692a
fix closing square bracket in figure legends
