### 2026-01-16 00a46c0
remove info to reproduce plots from main text
