### 2026-01-16 fb0b118
add point to legend about names being simulated in multi-dot-plot figure
