+++ joss/paper.md
@@ -49,7 +49,7 @@ However, to our knowledge, there are no existing tools that offer an easy method
 
 `simulist` offers two modules of functionality: 1) simulation and 2) post-processing. In the simulation module there is a function for simulating a line list, a function for simulating contact tracing data, and a function for simulating both. Respectively these output, an outbreak line list (`data.frame`), a contact tracing data set (`data.frame`), and lastly both an outbreak line list and contact tracing data set that are linked (i.e. from the same outbreak) (`list` of `data.frame`s). All simulation functions internally call a branching process model [@farringtonBranchingProcessModels2003a] with the option of network adjusted contact patterns to account for the fact that choosing someone at random by following up a contact chooses individuals with a probability proportional to their number of contacts [@alexanderRiskFactorsEvolutionary2010; @blumbergDetectingDifferentialTransmissibilities2014]. No data is stored in the package, all outbreak data is generated upon calling a simulation function. All simulation functions can be run without requiring the user to specify any parameters (i.e. function arguments, see function documentation for defaults), but all functions offer a range of arguments to modify the characteristics of the outbreak data. The parameters that can be specified by the user are organised into two levels. The parameters that control: the branching process, key epidemiological delays, risks, and demographic parameters are function arguments because they are more likely to be adjusted by users. Parameters thought to rarely require changing, except for advanced use cases are passed to the function via a configuration list, to prevent the function signature being overly complex. A helper function is provided to automatically generate the configuration list with sensible defaults, and allow easy modification, and this is called by default in the simulation functions.
 
-![A connected dot plot of the timeline of events in a simulated line list. See _Visualising simulated data vignette_ at <https://epiverse-trace.github.io/simulist/articles/vis-linelist.html> or in R (`vignette("vis-linelist", package = "simulist")`) for more information.\label{fig:events}](multi-dot-plot.png)
+![A connected dot plot of the timeline of events in a simulated line list. Data is simulated and all case names are randomly generated. See _Visualising simulated data vignette_ at <https://epiverse-trace.github.io/simulist/articles/vis-linelist.html> or in R (`vignette("vis-linelist", package = "simulist")`) for more information.\label{fig:events}](multi-dot-plot.png)
 
 ![The daily incidence of symptom onset, hospital admission and deaths facetted by sex for a simulated line list and aggregated and plotted with `incidence2` [@taylorIncidence2ComputeHandle2020]. See _Visualising simulated data vignette_ at <https://epiverse-trace.github.io/simulist/articles/vis-linelist.html> or in R (`vignette("vis-linelist", package = "simulist")`) for more information.\label{fig:incidence}](incidence.png) 
 
@@ -61,7 +61,7 @@ Here we highlight some of the simulation functionality available:
 
 ***Age-structured population***: the outbreak simulation in `simulist` assumes a uniform population age between a lower and upper age range (0 and 90 by default). However, in order to accurately generate synthetic data for outbreak scenarios where the population age structure is known this can be input into the outbreak simulation, again using a user-defined `data.frame` to flexibility specify an arbitrary number of age groups and proportions. See the [Age-structured population vignette](https://epiverse-trace.github.io/simulist/articles/age-struct-pop.html) for more details and a worked example.
 
[…]
