### 2026-01-16 f59b8be
Updates to JOSS manuscript from peer review (#274)
* try and fix \ref

* try \autoref with a single image

* try and trigger GHA

* remove figure legend

* try autoref with smaller fig width

* try with second plot

* split into three plots

* add links to vis-linelist vignette and figures script

* fix links in figure legends

* fix closing square bracket in figure legends

* remove info to reproduce plots from main text

* reorder legend in multi-dot-plot.png

* add point to legend about names being simulated in multi-dot-plot figure
