+++ joss/joss-figs.R
@@ -84,6 +84,11 @@ tidy_linelist <- linelist %>%
   mutate(ordering_value = ifelse(name == "date_onset", value, NA)) %>% # nolint consecutive_mutate_linter
   mutate(case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)) # nolint consecutive_mutate_linter
 
+tidy_linelist$name <- factor(
+  tidy_linelist$name,
+  levels = c("date_onset", "date_reporting", "date_admission", "date_outcome")
+)
+
 ll_plot <- ggplot(data = tidy_linelist) +
   geom_line(
     mapping = aes(x = value, y = case_name),
@@ -95,24 +100,32 @@ ll_plot <- ggplot(data = tidy_linelist) +
       y = case_name,
       shape = name,
       col = name
-    ), size = 2
+    ),
+    size = 2
   ) +
   scale_x_date(name = "Event date", date_breaks = "week") +
   scale_y_discrete(name = "Case name") +
   scale_color_brewer(
     palette = "Set1",
     name = "Event type",
-    labels = c("Date Admission", "Date Onset", "Date Outcome", "Date Reporting")
+    labels = c("Date Onset", "Date Reporting", "Date Admission", "Date Outcome")
   ) +
   scale_shape_manual(
     name = "Event type",
     labels = c(
-      "Date Admission", "Date Onset", "Date Outcome", "Date Reporting"
+      "Date Onset", "Date Reporting", "Date Admission", "Date Outcome"
     ),
     values = c(15, 16, 17, 18)
   ) +
   theme_bw() +
-  theme(legend.position = "bottom", axis.text.y = element_text(size = 4))
+  theme(
+    legend.position = "bottom",
+    axis.text.x = element_text(
+      angle = 45,
+      vjust = 1,
+      hjust = 1
+    )
+  )
 
 
 ggplot2::ggsave(
