### 2026-01-19 #8781 @fabian-s
thanks @joshwlambert for the well-structured and constructive response.

@mponce0 @gdenicola please take a look at the revised package & paper and tick off the respective open check list points if your concerns were adressed successfully. If not, please let us know if/which issues persist.
We would appreciate your responses **preferably before Feb 9**.
