+++ dev/articles/age-strat-risks.md
@@ -118,13 +118,13 @@ linelist <- sim_linelist(
 
 # first 6 rows of linelist
 head(linelist)
-#>   id              case_name case_type sex age date_onset date_reporting
-#> 1  1           Kevin Pullen suspected   m   1 2023-01-01     2023-01-01
-#> 2  2 Carisa Flores-Gonzalez confirmed   f  29 2023-01-01     2023-01-01
-#> 3  3       Maazin el-Othman confirmed   m  78 2023-01-01     2023-01-01
-#> 4  5       Faisal el-Vaziri suspected   m  70 2023-01-01     2023-01-01
-#> 5  6           Lynsey Duron confirmed   f  28 2023-01-01     2023-01-01
-#> 6  8         Lilibeth Black confirmed   f  61 2023-01-01     2023-01-01
+#>   id            case_name case_type sex age date_onset date_reporting
+#> 1  1   Jennifer Pritchett confirmed   f   1 2023-01-01     2023-01-01
+#> 2  2         Tyler Payson confirmed   f  29 2023-01-01     2023-01-01
+#> 3  3            Sean Wong confirmed   m  78 2023-01-01     2023-01-01
+#> 4  5       Bishr al-Safar confirmed   m  70 2023-01-01     2023-01-01
+#> 5  6 Francisco Montgomery  probable   m  28 2023-01-01     2023-01-01
+#> 6  8         Jack Millard suspected   m  61 2023-01-01     2023-01-01
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
 #> 1     2023-01-03      died   2023-01-18               <NA>              <NA>
 #> 2     2023-01-03      died   2023-02-09         2022-12-30        2023-01-08
@@ -133,12 +133,12 @@ head(linelist)
 #> 5     2023-01-05 recovered         <NA>         2022-12-29        2023-01-02
 #> 6           <NA> recovered         <NA>         2022-12-28        2023-01-05
 #>   ct_value
-#> 1       NA
-#> 2     25.8
-#> 3     24.9
-#> 4       NA
-#> 5     24.5
-#> 6     26.4
+#> 1     27.1
+#> 2     25.2
+#> 3     24.8
+#> 4     26.2
+#> 5       NA
+#> 6       NA
 ```
 
 We can run another simulation and change the hospitalisation and death
@@ -160,27 +160,27 @@ linelist <- sim_linelist(
 )
 
 head(linelist)
-#>   id          case_name case_type sex age date_onset date_reporting
-#> 1  1           Kacy Kim suspected   f  80 2023-01-01     2023-01-01
-#> 2  2        Jina Warnes  probable   f  85 2023-01-01     2023-01-01
-#> 3  4     Raadi el-Yasin  probable   m  76 2023-01-01     2023-01-01
-#> 4  8   Joshua Castaneda confirmed   m  12 2023-01-01     2023-01-01
-#> 5 11 Fat'hiyaa al-Zafar suspected   f  50 2023-01-01     2023-01-01
-#> 6 14    Matthew Sheldon  probable   m  54 2023-01-01     2023-01-01
+#>   id         case_name case_type sex age date_onset date_reporting
+#> 1  1    Karlena Garner confirmed   f  32 2023-01-01     2023-01-01
+#> 2  3 Cassandra Le Beau suspected   f  48 2023-01-01     2023-01-01
+#> 3  4      Lauren Moran confirmed   f   3 2023-01-01     2023-01-01
+#> 4  5       Alexis Ware confirmed   f  25 2023-01-01     2023-01-01
+#> 5  6       Sophia Cruz  probable   f  57 2023-01-01     2023-01-01
[…]
