+++ dev/articles/vis-linelist.md
@@ -201,6 +201,13 @@ outbreak increasing in incidence and then declining, with epidemic
 curves similar between
 sexes.](vis-linelist_files/figure-html/plot-group-by-sex-1.png)
 
+The simulated data plotted above used the default equal probability of
+each contact and case being male or female. To adjust the probability
+that each contact/case is male/female adjust `prob_male` in the `config`
+argument (see
+[`create_config()`](https://epiverse-trace.github.io/simulist/dev/reference/create_config.md)
+for details).
+
 To visualise the onset, hospitalisation and death incidence in the same
 plot they can be jointly specified to the `date_index` argument of
 [`incidence2::incidence()`](https://rdrr.io/pkg/incidence2/man/incidence.html).
@@ -331,13 +338,13 @@ outbreak <- sim_outbreak(
   onset_to_death = onset_to_death
 )
 head(outbreak$linelist)
-#>   id              case_name case_type sex age date_onset date_reporting
-#> 1  1           Kevin Pullen suspected   m   1 2023-01-01     2023-01-01
-#> 2  2 Carisa Flores-Gonzalez confirmed   f  29 2023-01-01     2023-01-01
-#> 3  3       Maazin el-Othman confirmed   m  78 2023-01-02     2023-01-02
-#> 4  5       Faisal el-Vaziri suspected   m  70 2023-01-05     2023-01-05
-#> 5  6           Lynsey Duron confirmed   f  28 2023-01-02     2023-01-02
-#> 6  8         Lilibeth Black confirmed   f  61 2023-01-04     2023-01-04
+#>   id            case_name case_type sex age date_onset date_reporting
+#> 1  1   Jennifer Pritchett confirmed   f   1 2023-01-01     2023-01-01
+#> 2  2         Tyler Payson confirmed   f  29 2023-01-01     2023-01-01
+#> 3  3            Sean Wong confirmed   m  78 2023-01-02     2023-01-02
+#> 4  5       Bishr al-Safar confirmed   m  70 2023-01-05     2023-01-05
+#> 5  6 Francisco Montgomery  probable   m  28 2023-01-02     2023-01-02
+#> 6  8         Jack Millard suspected   m  61 2023-01-04     2023-01-04
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
 #> 1     2023-01-03      died   2023-01-18               <NA>              <NA>
 #> 2     2023-01-03      died   2023-02-09         2022-12-30        2023-01-08
@@ -346,27 +353,27 @@ head(outbreak$linelist)
 #> 5     2023-01-06 recovered         <NA>         2022-12-30        2023-01-03
 #> 6           <NA> recovered         <NA>         2022-12-29        2023-01-06
 #>   ct_value
-#> 1       NA
-#> 2     25.8
-#> 3     24.9
-#> 4       NA
-#> 5     24.5
-#> 6     26.4
+#> 1     27.1
+#> 2     25.2
+#> 3     24.8
+#> 4     26.2
+#> 5       NA
+#> 6       NA
 head(outbreak$contacts)
-#>                     from                     to age sex date_first_contact
-#> 1           Kevin Pullen Carisa Flores-Gonzalez  29   f         2022-12-30
-#> 2           Kevin Pullen       Maazin el-Othman  78   m         2022-12-31
-#> 3 Carisa Flores-Gonzalez         Katelyn Catlin  22   f         2022-12-30
-#> 4 Carisa Flores-Gonzalez       Faisal el-Vaziri  70   m         2022-12-31
[…]
