### 2026-01-19 906d76b
add and call incubation_period in .sim_internal, relates #253
