### 2026-01-19 d1066b7
add nolint flags for explicit returns in .sim_internal
