### 2026-01-22 #8781 @gdenicola
Thanks @joshwlambert for the response and corresponding edits. I have updated the checklist and ticked the remaining boxes. The only outstanding issue I see is that the phrase I previously mentioned:

“The majority of outbreak analytics uses R (R Core Team, 2025) and R packages extending the language to answer common epidemiological questions in an outbreak (Tamayo Cuartero et al., 2025). From basic data wrangling and data science using tools such as the tidyverse (Wickham et al., 2019), to using advanced methods, for example nowcasting/forecasting of transmissibility (Abbott et al., 2020) This is accompanied by […]”

is still showing as it was in the previous version (with zero changes, perhaps a technical issue?), so it would be good to fix the syntax there. This is minor of course, and apart from that, the paper is good to go on my side. Thanks!
