### 2026-01-29 #8781 @joshwlambert
Thanks @gdenicola! I've check the paper both on the [{simulist} GitHub repo](https://github.com/epiverse-trace/simulist/blob/joss/joss/paper.md?plain=1#L44) and the [JOSS proof](https://github.com/openjournals/joss-papers/blob/joss.08781/joss.08781/10.21105.joss.08781.pdf) and it looks like the paragraph "The majority of ..." has been updated. It may have been a technical issue, if not please let me know and I'll update.
