### 2026-01-29 #8781 @mponce0
@joshwlambert Thank you for the update and great work you and your team have put into the simulist package.
In particular, we appreciate the effort and investment in the creation of vignettes and tutorials associated to the package in epiverse.

There is only one minor point for me, which related to the "license" files present in the repo: currently there are two files -- `LICENSE` and `LICENSE.md`.  I'm not sure whether the former file is needed for some technical requirements but if not, then  I'd recommend to remove it for clarity sake.
If you could confirm or comment about that, then that would be all from my side.
