### 2026-01-29 #8781 @joshwlambert
Thanks for the kind words @mponce0. On the point around the license files:

There are two license files in the R package: `LICENSE` and `LICENSE.md`. This is the standard format for licensing in R packages and follows the Comprehensive R Archive Network (CRAN) [policy on licenses](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing-1), as well as the the [Licensing chapter in the R packages book](https://r-pkgs.org/license.html), where the pertinent information states: "`LICENSE.md` includes a copy of the full text of the license. All open source licenses require a copy of the license to be included, but CRAN does not permit you to include a copy of standard licenses in your package, so we also use `.Rbuildignore` to make sure this file is not sent to CRAN."

This is fairly standard practise for R packages and is common in widely used R packages (e.g. tidyverse packages).
