### 2026-01-29 epiverse-trace/simulist#276 @fabian-s
JOSS paper: typos, language etc
https://github.com/epiverse-trace/simulist/blob/f59b8bec5aac1dfb2111b8d72da72a07c5a4ff10/joss/paper.md?plain=1#L44

Please add a period after `[@abbottEstimatingTimevaryingReproduction2020a]`

https://github.com/epiverse-trace/simulist/blob/f59b8bec5aac1dfb2111b8d72da72a07c5a4ff10/joss/paper.md?plain=1#L46

- compl**e**mented, not compl**i**mented
- "messying": pretty sure that is *not* a real word ... ;)
- " ... types of issues encounter**ED** ...."

please get a native speaker (or an AI) to proofread this properly, it's not ready for publication as-is.

#https://github.com/openjournals/joss-reviews/issues/8781
