### 2026-01-29 0578087
fix spelling and grammar in JOSS paper, closes #276
