### 2026-01-29 59606cf
fix spelling and grammar in JOSS paper, closes #276
