### 2026-02-02 #8781 @joshwlambert
@fabian-s please let me know if there is anything else required from my side and when I should make the new version release of {simulist} to accompany the JOSS paper.
