### 2026-02-02 epiverse-trace/simulist#276 @fabian-s
Closed this too early, sorry --

Grammatical errors / typos:

 - Line ~57: "From basic data wrangling..." is a sentence fragment. It starts with "From..." but has no main clause — the period after the previous sentence should be a comma or semicolon, or this fragment needs a verb. Consider: "These range from basic data wrangling...to using advanced methods..."
  - Line ~79: "facetted" should be "faceted" (standard spelling)
  - Line ~89: "to flexibility specify" should be "to flexibly specify" (adverb, not noun)
  - Line ~71: "A helper function is provided to automatically generate the configuration list with sensible defaults, and allow easy modification, and this is called by default" — awkward double "and". Consider: "...with sensible defaults and allow easy modification; this helper is called by default..."

  Style preferences

  - Line ~35: "preparedness and analytics response" reads awkwardly. Consider "preparedness and analytical response" or "preparedness, analytics, and response"
  - Line ~35: "in particular, infectious disease epidemiology (e.g. delay distributions)" reads as if epidemiology itself is a variable being parameterised. Consider restructuring: "— including epidemiological parameters (e.g. delay distributions), contact patterns, and demographic information —"
  - Line ~44: "adequacy" is vague in "testing analysis methods for precision, adequacy and robustness". Consider "accuracy" or "validity"
  - Line ~71: The sentence starting "All simulation functions internally call a branching process model..." is very long. Consider splitting after "...contact patterns"
  - Line ~75: "Data is simulated" should be "Data are simulated" (conventionally plural in academic writing)
  - Line ~85: "which all, by default accept" should be "which all, by default**,** accept" (closing comma around parenthetical)
  - Line ~85: "so any age-stratification is accepted" is unclear. Consider: "so that any age-stratification can be specified"
  - Line ~89: "in order to accurately generate" could be simplified to "to accurately generate"
  - Line ~99: "However, in reality those working" should be "However, in reality**,** those working" (comma after introductory phrase)
  - Line ~109: "and help developer onboarding" should be "and help with developer onboarding" or "and facilitate developer onboarding"
  - Line ~113: "it is also included within" — this is a comma splice. Consider: "and it is also included within" or use a semicolon

(reviewed by Opus 4.5, then reviewed by me - please do use the tools we have available these days to produce high-quality work...)
