### 2026-02-02 c23c335
fix spelling and grammar issues, closes #276
