+++ joss/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 ![simulist package hex logo, containing the words simulist and underneath, powered by data.org, with overlapping rectangles above the words](../man/figures/logo.svg){ width=240px }\
 
-Epidemic and pandemic preparedness and analytics response requires robust analysis methods and a deep understanding of outbreak data. We introduce `simulist`, an open-source R package for simulating realistic infectious disease outbreak data. It is designed to allow users to generate data with specific disease outbreak characteristics by enabling flexible parameterisation of processes and variables, in particular, infectious disease epidemiology (e.g. delay distributions), contact patterns and demographic information, to simulate two common datasets in outbreak settings: 1) line list data, and 2) contact tracing data. It also offers post-processing of line list data to replicate right-truncation of real-time outbreak data, as well as creating "messy" data with realistically incomplete and inconsistent values. 
+Epidemic and pandemic preparedness and response requires robust analysis methods and a deep understanding of outbreak data. We introduce `simulist`, an open-source R package for simulating realistic infectious disease outbreak data. It is designed to allow users to generate data with specific disease outbreak characteristics by enabling flexible parameterisation of processes and variables, in particular, infectious disease epidemiology (e.g. delay distributions), contact patterns and demographic information, to simulate two common datasets in outbreak settings: 1) line list data, and 2) contact tracing data. It also offers post-processing of line list data to replicate right-truncation of real-time outbreak data, as well as creating "messy" data with realistically incomplete and inconsistent values. 
 
 ## Statement of need
 
@@ -41,15 +41,15 @@ In the ecosystem of R packages for epidemiology, there are a range of tools for 
 
 ## State of the field
 
-The majority of outbreak analytics uses R [@rcoreteamLanguageEnvironmentStatistical2025]. There are many R packages that help to answer common epidemiological questions in an outbreak [@tamayocuartero100DayMission2025]. From basic data wrangling and data science tools such as the tidyverse [@wickhamWelcomeTidyverse2019], to using advanced methods, for example nowcasting/forecasting of transmissibility [@abbottEstimatingTimevaryingReproduction2020a]. Additionally, there are several initiatives that teach epidemiology using R tooling [@batraEpidemiologistHandbook2021; @vallecamposEpiverseTRACETutorials2025], resulting in an active ecosystem of open-source tooling, teaching, and community engagement [@morganWHOHubPandemic2022], for example the [epinowcast community](https://www.epinowcast.org/).
+The majority of outbreak analytics uses R [@rcoreteamLanguageEnvironmentStatistical2025]. There are many R packages that help to answer common epidemiological questions in an outbreak [@tamayocuartero100DayMission2025]. These range from basic data wrangling and data science tools such as the tidyverse [@wickhamWelcomeTidyverse2019], to using advanced methods, for example nowcasting/forecasting of transmissibility [@abbottEstimatingTimevaryingReproduction2020a]. Additionally, there are several initiatives that teach epidemiology using R tooling [@batraEpidemiologistHandbook2021; @vallecamposEpiverseTRACETutorials2025], resulting in an active ecosystem of open-source tooling, teaching, and community engagement [@morganWHOHubPandemic2022], for example the [epinowcast community](https://www.epinowcast.org/).
 
[…]
