### 2026-02-02 b5302a8
fix spelling and grammar issues, closes #276
