### 2026-02-02 cfe5d74
clarify wording on epidemiological parameterisation
