### 2026-02-02 4a3e094
split sentence on risk stratification
