+++ joss/paper.md
@@ -57,7 +57,7 @@ However, to our knowledge, there are no existing tools that offer an easy method
 
 Here we highlight some of the simulation functionality available:
 
-***Age-stratified hospitalisation and death risks***: the outbreak simulation includes three risks: hospitalisation risk, death risk in hospital, and death risk outside of hospital, which all, by default, accept a single number to define the population risk. In many infectious disease outbreaks risk is non-uniform across ages, so `simulist` accepts risks stratified by age using a user-defined `data.frame`, so that any age-stratification can be specified. See the [Age-stratified hospitalisation and death risks vignette](https://epiverse-trace.github.io/simulist/index.html) for a complete explanation.
+***Age-stratified hospitalisation and death risks***: the outbreak simulation includes three risks: hospitalisation risk, death risk in hospital, and death risk outside of hospital, which all, by default, accept a single number to define the population risk. In many infectious disease outbreaks risk is non-uniform across ages. `simulist` accepts risks stratified by age using a user-defined `data.frame`, so that any age-stratification can be specified. See the [Age-stratified hospitalisation and death risks vignette](https://epiverse-trace.github.io/simulist/index.html) for a complete explanation.
 
 ***Age-structured population***: the outbreak simulation in `simulist` assumes a uniform population age between a lower and upper age range (0 and 90 by default). However, to accurately generate synthetic data for outbreak scenarios where the population age structure is known this can be input into the outbreak simulation, again using a user-defined `data.frame` to flexibly specify an arbitrary number of age groups and proportions. See the [Age-structured population vignette](https://epiverse-trace.github.io/simulist/articles/age-struct-pop.html) for more details and a worked example.
 
