### 2026-02-03 #8781 @fabian-s
@joshwlambert please see points above
