### 2026-02-03 #8781 @fabian-s
@joshwlambert
- link in "See the Age-stratified hospitalisation and death risks vignette for a complete explanation." points to https://epiverse-trace.github.io/simulist/index.html and not the vignette, please fix that as well.
- please check the missing DOI the bot found
