### 2026-02-03 e90fca1
fix links in JOSS paper
