### 2026-02-03 ff5100e
add Visualising line list events section to vis-linelist.Rmd
