+++ vignettes/vis-linelist.Rmd
@@ -25,6 +25,9 @@ library(simulist)
 library(epiparameter)
 library(incidence2)
 library(epicontacts)
+library(tidyr)
+library(dplyr)
+library(ggplot2)
 ```
 
 First we load the required delay distributions using the {epiparameter} package.
@@ -193,8 +196,6 @@ linelist$date_recovery[linelist$outcome == "died"] <- NA
 #### Tidyverse
 
 ```{r, reshape-linelist-tidyverse, message=FALSE}
-library(tidyr)
-library(dplyr)
 linelist <- linelist %>%
   tidyr::pivot_wider(
     names_from = outcome,
@@ -231,6 +232,80 @@ daily <- incidence(
 #|   (rows). The plot shows the outbreak fluctuating, with many fewer
 #|   hospital admissions and deaths than cases.
 plot(daily)
+```
+
+## Visualising line list events through time
+
+Instead of plotting the number of cases on a particular day, we can plot each individual's events over the course of their infection.
+
+We start by simulating a line list. We include an onset-to-recovery delay distribution so all cases have an outcome, a reporting delay so the date cases are reported is after their symptom onset, and set a high risk of hospitalisation to get many hospital admission events. We subset to the first 10 cases so the events on the plot are clear, however, this subsetting is not required.
+
+```{r, sim-subset-linelist}
+set.seed(123)
+linelist <- sim_linelist(
+  onset_to_recovery = function(x) stats::rlnorm(n = x, meanlog = 2, sdlog = 0.5),
+  reporting_delay = function(x) stats::rlnorm(n = x, meanlog = 1, sdlog = 0.5),
+  hosp_risk = 0.8
+)
+linelist <- linelist[1:10, ]
+```
+
+We need to reshape the line list to _tidy_ data in order to easily plot it with {ggplot2}.
+
+```{r, prep-events}
+tidy_linelist <- linelist %>%
+  pivot_longer(
+    cols = c("date_onset", "date_reporting", "date_admission", "date_outcome")
+  ) %>%
+  mutate(
+    ordering_value = ifelse(name == "date_onset", value, NA),
+    case_name = reorder(case_name, ordering_value, min, na.rm = TRUE)
+) 
+  
+tidy_linelist$name <- factor(
+  tidy_linelist$name,
+  levels = c("date_onset", "date_reporting", "date_admission", "date_outcome")
+)
+```
+
+```{r, plot-events}
+ggplot(data = tidy_linelist) +
+  geom_line(
+    mapping = aes(x = value, y = case_name),
+    linewidth = 0.25
+  ) +
+  geom_point(
+    mapping = aes(
+      x = value,
+      y = case_name,
+      shape = name,
+      col = name
+    ),
+    size = 2
+  ) +
+  scale_x_date(name = "Event date", date_breaks = "week") +
+  scale_y_discrete(name = "Case name") +
+  scale_color_brewer(
+    palette = "Set1",
+    name = "Event type",
+    labels = c("Date Onset", "Date Reporting", "Date Admission", "Date Outcome")
+  ) +
+  scale_shape_manual(
+    name = "Event type",
+    labels = c(
+      "Date Onset", "Date Reporting", "Date Admission", "Date Outcome"
+    ),
+    values = c(15, 16, 17, 18)
+  ) +
+  theme_bw() +
+  theme(
+    legend.position = "bottom",
+    axis.text.x = element_text(
+      angle = 45,
+      vjust = 1,
+      hjust = 1
+    )
+  )
 ```
 
 ## Demographic data
[…]
