### 2026-02-03 e9e33f2
use <epiparameter> objects in Visualising line list events section in vis-linelist.Rmd
