### 2026-02-03 #8781 @joshwlambert
@fabian-s I've reviewed the bibliography and the DOI check, the missing DOI is not an official DOI for the R language, but rather for its help pages. Therefore I'm happy that all the DOIs that are available for references are in `paper.bib`.

Thanks for spotting the wrong link, apologies for not catching this sooner, I've fixed this (epiverse-trace/simulist@d0e2c395e3df3be229bbcdd0932de08fe436ea09) and checked all other links.
