### 2026-02-04 e8ba32a
update NEWS for v0.7.0 release
