### 2026-02-05 41f6bf9
update .Rbuildignore and paths to paper folder
