### 2026-02-05 0b539e6
update .Rbuildignore and paths to paper folder
