+++ README.md
@@ -96,13 +96,13 @@ data set.
 set.seed(1)
 linelist <- sim_linelist()
 head(linelist)
-#>   id        case_name case_type sex age date_onset date_reporting
-#> 1  1 Lolette Phillips suspected   f  59 2023-01-01     2023-01-01
-#> 2  2       James Jack suspected   m  90 2023-01-01     2023-01-01
-#> 3  3      Chen Kantha confirmed   m   4 2023-01-02     2023-01-02
-#> 4  5  Saleema al-Zaki  probable   f  29 2023-01-04     2023-01-04
-#> 5  6     David Ponzio confirmed   m  14 2023-01-05     2023-01-05
-#> 6  7 Christopher Ward  probable   m  85 2023-01-06     2023-01-06
+#>   id          case_name case_type sex age date_onset date_reporting
+#> 1  1        James Manis suspected   m  59 2023-01-01     2023-01-01
+#> 2  2          Chen Moua confirmed   m  90 2023-01-01     2023-01-01
+#> 3  3       David Welter confirmed   m   4 2023-01-02     2023-01-02
+#> 4  5 Christopher Turner confirmed   m  29 2023-01-04     2023-01-04
+#> 5  6        Morgan Bohn suspected   f  14 2023-01-05     2023-01-05
+#> 6  7    Yutitham Corpuz  probable   m  85 2023-01-06     2023-01-06
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
 #> 1     2023-01-09      died   2023-01-13               <NA>              <NA>
 #> 2           <NA> recovered         <NA>         2022-12-29        2023-01-03
@@ -112,10 +112,10 @@ head(linelist)
 #> 6     2023-01-08 recovered         <NA>         2022-12-31        2023-01-06
 #>   ct_value
 #> 1       NA
-#> 2       NA
+#> 2     24.5
 #> 3     24.8
-#> 4       NA
-#> 5     24.6
+#> 4     25.4
+#> 5       NA
 #> 6       NA
 ```
 
@@ -211,13 +211,13 @@ linelist <- sim_linelist(
   onset_to_death = onset_to_death
 )
 head(linelist)
-#>   id              case_name case_type sex age date_onset date_reporting
-#> 1  1           Kevin Pullen suspected   m   1 2023-01-01     2023-01-01
-#> 2  2 Carisa Flores-Gonzalez confirmed   f  29 2023-01-01     2023-01-01
-#> 3  3       Maazin el-Othman confirmed   m  78 2023-01-01     2023-01-01
-#> 4  5       Faisal el-Vaziri suspected   m  70 2023-01-01     2023-01-01
-#> 5  6           Lynsey Duron confirmed   f  28 2023-01-01     2023-01-01
-#> 6  8         Lilibeth Black confirmed   f  61 2023-01-01     2023-01-01
+#>   id            case_name case_type sex age date_onset date_reporting
+#> 1  1   Jennifer Pritchett confirmed   f   1 2023-01-01     2023-01-01
+#> 2  2         Tyler Payson confirmed   f  29 2023-01-01     2023-01-01
+#> 3  3            Sean Wong confirmed   m  78 2023-01-01     2023-01-01
+#> 4  5       Bishr al-Safar confirmed   m  70 2023-01-01     2023-01-01
+#> 5  6 Francisco Montgomery  probable   m  28 2023-01-01     2023-01-01
+#> 6  8         Jack Millard suspected   m  61 2023-01-01     2023-01-01
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
 #> 1     2023-01-03      died   2023-01-18               <NA>              <NA>
 #> 2     2023-01-03      died   2023-02-09         2022-12-30        2023-01-08
[…]
