+++ cran-comments.md
@@ -1,7 +1,5 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 0 notes
+0 errors | 0 warnings | 1 note
 
 * Running R CMD check (with _R_CHECK_INCOMING_REMOTE_) flags 403 warnings for LSHTM URLs but these are valid URLs and are not flagged on winbuilder.
-
-* There are no papers describing the methods used in this package.
