+++ articles/simulist.md
@@ -104,13 +104,13 @@ linelist <- sim_linelist(
   onset_to_death = onset_to_death
 )
 head(linelist)
-#>   id             case_name case_type sex age date_onset date_reporting
-#> 1  1         Fabian Mrazik confirmed   m  90 2023-01-01     2023-01-01
-#> 2  3       Ashley Martinez confirmed   f  71 2023-01-01     2023-01-01
-#> 3  4                Tia Vu  probable   f  48 2023-01-01     2023-01-01
-#> 4  5 Abdul Majeed el-Saleh confirmed   m  77 2023-01-01     2023-01-01
-#> 5  6        Courtney Flood suspected   f  83 2023-01-02     2023-01-02
-#> 6  7          Joseph Jiron suspected   m  56 2023-01-01     2023-01-01
+#>   id                case_name case_type sex age date_onset date_reporting
+#> 1  1            Joseph Lovato suspected   m  90 2023-01-01     2023-01-01
+#> 2  3         Rayyaan al-Basha  probable   m  71 2023-01-01     2023-01-01
+#> 3  4                Tara Yang  probable   f  48 2023-01-01     2023-01-01
+#> 4  5           Brianna Madden confirmed   f  77 2023-01-01     2023-01-01
+#> 5  6 Miguel Cabral-Hollowhorn suspected   m  83 2023-01-02     2023-01-02
+#> 6  7           Matthew Samadh confirmed   m  56 2023-01-01     2023-01-01
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
 #> 1           <NA> recovered         <NA>               <NA>              <NA>
 #> 2     2023-01-04      died   2023-01-10         2022-12-26        2023-01-06
@@ -119,12 +119,12 @@ head(linelist)
 #> 5           <NA> recovered         <NA>         2022-12-26        2023-01-04
 #> 6           <NA> recovered         <NA>         2022-12-28        2023-01-03
 #>   ct_value
-#> 1     21.9
-#> 2     22.7
+#> 1       NA
+#> 2       NA
 #> 3       NA
-#> 4     27.4
+#> 4     24.1
 #> 5       NA
-#> 6       NA
+#> 6     26.7
 ```
 
 ## Controlling outbreak size
@@ -195,13 +195,13 @@ linelist <- sim_linelist(
   outbreak_size = c(250, 1e4)
 )
 head(linelist)
-#>   id            case_name case_type sex age date_onset date_reporting
-#> 1  1          John Farley confirmed   m  61 2023-01-01     2023-01-01
-#> 2  2      Jared Weixelman confirmed   m  29 2023-01-01     2023-01-01
-#> 3  3         Kyla Collins confirmed   f  71 2023-01-01     2023-01-01
-#> 4  4    Shianne Crawleigh  probable   f  23 2023-01-01     2023-01-01
-#> 5  5 Kathryn Pfannenstiel confirmed   f   7 2023-01-01     2023-01-01
-#> 6  6     Catherine Morris  probable   f  19 2023-01-01     2023-01-01
+#>   id          case_name case_type sex age date_onset date_reporting
+#> 1  1  Naaila el-Barakat  probable   f  61 2023-01-01     2023-01-01
+#> 2  2       Kevin Herman suspected   m  29 2023-01-01     2023-01-01
+#> 3  3 Sameera el-Soliman confirmed   f  71 2023-01-01     2023-01-01
+#> 4  4        Tate Barrow confirmed   m  23 2023-01-01     2023-01-01
+#> 5  5     Brandi Goldman confirmed   f   7 2023-01-01     2023-01-01
+#> 6  6      Faith Berrien suspected   f  19 2023-01-01     2023-01-01
 #>   date_admission   outcome date_outcome date_first_contact date_last_contact
[…]
