+++ articles/vis-linelist.md
@@ -16,7 +16,19 @@ library(simulist)
 library(epiparameter)
 library(incidence2)
 #> Loading required package: grates
+library(ggplot2)
 library(epicontacts)
+library(tidyr)
+library(dplyr)
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
+library(ggplot2)
 ```
 
 First we load the required delay distributions using the {epiparameter}
@@ -201,6 +213,13 @@ outbreak increasing in incidence and then declining, with epidemic
 curves similar between
 sexes.](vis-linelist_files/figure-html/plot-group-by-sex-1.png)
 
+The simulated data plotted above used the default equal probability of
+each contact and case being male or female. To adjust the probability
+that each contact/case is male/female adjust `prob_male` in the `config`
+argument (see
+[`create_config()`](https://epiverse-trace.github.io/simulist/reference/create_config.md)
+for details).
+
 To visualise the onset, hospitalisation and death incidence in the same
 plot they can be jointly specified to the `date_index` argument of
 [`incidence2::incidence()`](https://rdrr.io/pkg/incidence2/man/incidence.html).
@@ -224,13 +243,11 @@ linelist$date_recovery[linelist$outcome == "died"] <- NA
 ```
 
 ``` r
-library(tidyr)
-library(dplyr)
-linelist <- linelist %>%
+linelist <- linelist |>
   tidyr::pivot_wider(
     names_from = outcome,
     values_from = date_outcome
-  ) %>%
+  ) |>
   dplyr::rename(
     date_death = died,
     date_recovery = recovered
@@ -264,6 +281,126 @@ data is facetted by sex (left female, right male) and incidence event
 (rows). The plot shows the outbreak fluctuating, with many fewer
 hospital admissions and deaths than
 cases.](vis-linelist_files/figure-html/plot-onset-hospitalisation-1.png)
+
+## Visualising individual line list events through time
+
+Instead of plotting the aggregated number of cases, deaths, or other
+events on a particular day, we can plot each individual’s timeline
+events over the course of their infection.
+
+We start by simulating a line list. We include an onset-to-recovery
+delay distribution so all cases have an outcome (died or recovered)
+date, a reporting delay so the date cases are reported is after their
+symptom onset date, and set a high risk of hospitalisation to get many
+hospital admission events. We subset to the first 10 cases so the events
+on the plot are clear, however, this subsetting is not required.
+
+``` r
+set.seed(123)
+
+onset_to_recovery <- epiparameter(
+  disease = "COVID-19",
+  epi_name = "onset to recovert",
+  prob_distribution = create_prob_distribution(
+    prob_distribution = "lnorm",
+    prob_distribution_params = c(meanlog = 2, sdlog = 0.5)
+  )
+)
+#> Citation cannot be created as author, year, journal or title is missing
+
+reporting_delay <- epiparameter(
+  disease = "COVID-19",
+  epi_name = "reporting delay",
+  prob_distribution = create_prob_distribution(
+    prob_distribution = "lnorm",
+    prob_distribution_params = c(meanlog = 1, sdlog = 0.5)
[…]
