### 2026-02-06 120b516
Update `CITATION.cff`
