+++ dev/articles/vis-linelist.md
@@ -1,721 +1,0 @@
-# Visualising simulated data
-
-This vignette gives an overview of the ways to plot the line list and
-contacts data output from the
-[`sim_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_linelist.md)
-and
-[`sim_outbreak()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_outbreak.md)
-functions.
-
-Plotting can be useful to identify certain transmission dynamics and
-patterns in the simulated data, or just to check that the simulated data
-looks as expected given how the simulation was parameterised.
-
-``` r
-library(simulist)
-library(epiparameter)
-library(incidence2)
-#> Loading required package: grates
-library(ggplot2)
-library(epicontacts)
-library(tidyr)
-library(dplyr)
-#> 
-#> Attaching package: 'dplyr'
-#> The following objects are masked from 'package:stats':
-#> 
-#>     filter, lag
-#> The following objects are masked from 'package:base':
-#> 
-#>     intersect, setdiff, setequal, union
-library(ggplot2)
-```
-
-First we load the required delay distributions using the {epiparameter}
-package.
-
-``` r
-contact_distribution <- epiparameter(
-  disease = "COVID-19",
-  epi_name = "contact distribution",
-  prob_distribution = create_prob_distribution(
-    prob_distribution = "pois",
-    prob_distribution_params = c(mean = 3)
-  )
-)
-#> Citation cannot be created as author, year, journal or title is missing
-
-infectious_period <- epiparameter(
-  disease = "COVID-19",
-  epi_name = "infectious period",
-  prob_distribution = create_prob_distribution(
-    prob_distribution = "gamma",
-    prob_distribution_params = c(shape = 3, scale = 2)
-  )
-)
-#> Citation cannot be created as author, year, journal or title is missing
-
-onset_to_hosp <- epiparameter(
-  disease = "COVID-19",
-  epi_name = "onset to hospitalisation",
-  prob_distribution = create_prob_distribution(
-    prob_distribution = "lnorm",
-    prob_distribution_params = c(meanlog = 1, sdlog = 0.5)
-  )
-)
-#> Citation cannot be created as author, year, journal or title is missing
-
-# get onset to death from {epiparameter} database
-onset_to_death <- epiparameter_db(
-  disease = "COVID-19",
-  epi_name = "onset to death",
-  single_epiparameter = TRUE
-)
-#> Using Linton N, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov A, Jung S, Yuan
-#> B, Kinoshita R, Nishiura H (2020). "Incubation Period and Other
-#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
-#> with Right Truncation: A Statistical Analysis of Publicly Available
-#> Case Data." _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
-#> <https://doi.org/10.3390/jcm9020538>.. 
-#> To retrieve the citation use the 'get_citation' function
-```
-
-Setting the seed ensures we have the same output each time the vignette
-is rendered. When using {simulist}, setting the seed is not required
-unless you need to simulate the same line list multiple times.
-
-``` r
-set.seed(123)
-```
-
-Using a simple line list simulation with the factory default settings:
-
-``` r
-linelist <- sim_linelist(
-  contact_distribution = contact_distribution,
-  infectious_period = infectious_period,
-  prob_infection = 0.33,
-  onset_to_hosp = onset_to_hosp,
-  onset_to_death = onset_to_death,
-  outbreak_size = c(500, 1e4)
-)
-```
-
-This line list contains 2066 cases.
-
-## Visualising incidence of onset, hospitalisation and death
-
[…]
