+++ dev/reference/dot-add_missing.md
@@ -1,26 +1,0 @@
-# Introduce user-specified proportion of custom missing values into a `<data.frame>`
-
-Introduce user-specified proportion of custom missing values into a
-`<data.frame>`
-
-## Usage
-
-``` r
-.add_missing(linelist, .args)
-```
-
-## Arguments
-
-- linelist:
-
-  Line list `<data.frame>` output from
-  [`sim_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/sim_linelist.md).
-
-- .args:
-
-  A list of setting from
-  [`messy_linelist()`](https://epiverse-trace.github.io/simulist/dev/reference/messy_linelist.md).
-
-## Value
-
-A line list `<data.frame>`

[…]
