### 2026-02-06 #8781 @joshwlambert
> Double check authors and affiliations (including ORCIDs)

✅

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

{simulist} v0.7.0 is the version coupled to the JOSS publication. It is released on [CRAN](https://cran.r-project.org/package=simulist) and [GitHub](https://github.com/epiverse-trace/simulist/releases/tag/v0.7.0).

> Archive the release on Zenodo/figshare/etc and post the DOI here.

{simulist} v0.7.0 has been archived in Zenodo. The version specific DOI is: [10.5281/zenodo.18505591](https://doi.org/10.5281/zenodo.18505591). There is also a version-agnostic DOI for all {simulist} releases which is: [10.5281/zenodo.10471458](https://doi.org/10.5281/zenodo.10471458).

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The ORCIDs for paper authors have been checked. They match the ORCIDs used in the `DESCRIPTION` file in the R package. The author list for the paper is not the same as the contributors listed in the `DESCRIPTION` file, all contributors listed as authors are also an author of the paper, additionally one paper author is listed as a package reviewer (`rev`) as they were the senior lead on the project.

> Make sure that the license listed for the archive is the same as the software license.

The license in the package is MIT, this matches the license on [Zenodo](https://zenodo.org/records/18505591).
