### 2026-02-06 #8781 @joshwlambert
The JOSS paper is now on the `main` branch of the {simulist} repository. The path to the paper (and bibliography, etc.) has changed from `joss/` to `paper/`. @fabian-s please let me know if you'd like me to update the information in this issue thread.
