### 2026-02-06 #8781 @fabian-s
@joshwlambert : only 2 authors listed at https://zenodo.org/records/18505591, paper has 3, please fix (and please be a little more detail-oriented, this kind of back-and-forth is just wasting all our time, frankly...)
