### 2026-02-06 #8781 @joshwlambert
> only 2 authors listed at https://zenodo.org/records/18505591, paper has 3, please fix

That is because of the way Zenodo automatically uses the authors from the `DESCRIPTION` file  in the R package. The reasons for the difference is outlined above (and pasted below):

> The ORCIDs for paper authors have been checked. They match the ORCIDs used in the DESCRIPTION file in the R package. The author list for the paper is not the same as the contributors listed in the DESCRIPTION file, all contributors listed as authors are also an author of the paper, additionally one paper author is listed as a package reviewer (rev) as they were the senior lead on the project.

Let me know if this is going to be a problem for proceeding with publication. Ideally I do not want to change the authors in the package as their current roles reflect their contributions to the project.
