### 2026-02-06 #8781 @joshwlambert
I've reinstated the `joss` branch on the GitHub repository and updated to (hopefully) have all of the updates from the `main` branch. The paper is in the `paper/` folder.
