### 2026-02-06 #8781 @joshwlambert
The first comment in this issue states: **Version:** v0.5.0, but the version with this paper is ~~v0.7.0~~ v0.7.1.
