### 2026-02-06 ede583d
update URL for reproducing figures, not using permalink
