### 2026-02-06 #8781 @fabian-s
> Let me know if this is going to be a problem for proceeding with publication. Ideally I do not want to change the authors in the package as their current roles reflect their contributions to the project.

It is, because the zenodo deposit is supposed to be the archival version of the JOSS paper (+ code), basically, so it makes no sense for it to have different authors than the paper.
Manually edit the author list for the zenodo deposit please ...?
