### 2026-02-06 3254372
fix typo in figure 2 legend
