### 2026-02-06 e6a29db
minor fixes to JOSS paper.md
