### 2026-02-06 1c5fcd7
update DESCRIPTION and NEWS.md to match main branch
