+++ NEWS.md
@@ -1,3 +1,7 @@
+# simulist 0.7.1
+
+* Minor fixes to _simulist_ JOSS paper for publication.
+
 # simulist 0.7.0
 
 This minor version release contains a range of minor updates and accompanies the _simulist_ publication in the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/).
