### 2026-02-06 9dad517
Update `CITATION.cff`
