### 2026-02-09 #8781 @joshwlambert
{simulist} v0.7.1 is the version release coupled with the JOSS submission. The archive and DOI is here: https://zenodo.org/records/18509079. I've updated the archive to match the JOSS author list.
