### 2026-02-09 #8781 @fabian-s
NB: missing reference above is a false positive
