### 2026-02-09 #8781 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
