### 2026-02-25 #8781 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the delay in getting this process started!). Below are some final checks, some of which may require your attention:
- [x] Please update the title of the archive to match that of the paper
### Copyediting stuff
- [x] first paragraph in State of the Field, second sentence: broken parallelism, suggest removing "using"
- [x] under Key functions: hyphenate "network-adjusted"
- [x] later the same paragraph: remove colon after "control"
- [x] next sentence: add comma after "advanced use cases"
- [x] footnote on first page continues on second in an awkward/unclear way – I bet if you slightly increase or reduce the size of the package logo it will either bump the whole thing onto the next page or allow the footnote to fit on the same page
- [x] in the list of functionalities beginning of page 5: capitalize the start of the sentences after the first two headers
- [x] in first functionality paragraph: add comma after "many infectious disease outbreaks"
- [ ] second functionality paragraph: suggest replacing "age range" with "limit"
- [ ] same paragraph: add comma after "is known"
- [ ] fourth functionality paragraph: add comma after "In `simulist`" and also after "Currently"
- [ ] under Package design: in first parenthetical, suggest changing to something like "(e.g. `epicontacts`, see Figure 3)"
- [ ] under Other resources: broken parallelism, suggest rephrasing to "`simulist` is documented through vignettes and function documentation, and is also included within Epiverse-TRACE open-source tutorials."
### Citations
- [ ] Please check the author list for the Epidemiologist R Handbook: some of the authors in the list appear to be usernames, but there's a list of names [here](https://www.epirhandbook.com/en/#contributors) you should be able to refer to
- [ ] I'm not sure that the data.org entry really needs to be a citation; probably an inline URL where you currently cite it would suffice since there doesn't seem to be a "formal" author list or DOI or anything
