### 2026-03-05 #8781 @joshwlambert
Thank you @rkurchin for taking the time to review the work.

> Please update the title of the archive to match that of the paper

The archive title has been updated to match the paper title, see https://zenodo.org/records/18509079.

> first paragraph in State of the Field, second sentence: broken parallelism, suggest removing "using"

I think this was in the third sentence, have removed "using".

> under Key functions: hyphenate "network-adjusted"

Done

> later the same paragraph: remove colon after "control"

Done

> next sentence: add comma after "advanced use cases"

Done

> footnote on first page continues on second in an awkward/unclear way – I bet if you slightly increase or reduce the size of the package logo it will either bump the whole thing onto the next page or allow the footnote to fit on the same page

I've reduced the width of the package logo from 240px to 200px.

> in the list of functionalities beginning of page 5: capitalize the start of the sentences after the first two headers

Done

> in first functionality paragraph: add comma after "many infectious disease outbreaks"

Done

> second functionality paragraph: suggest replacing "age range" with "limit"

Done

> same paragraph: add comma after "is known"

Done

> fourth functionality paragraph: add comma after "In simulist" and also after "Currently"

Done

> under Package design: in first parenthetical, suggest changing to something like "(e.g. epicontacts, see Figure 3)"

Done

> under Other resources: broken parallelism, suggest rephrasing to "simulist is documented through vignettes and function documentation, and is also included within Epiverse-TRACE open-source tutorials."

Done

> Please check the author list for the Epidemiologist R Handbook: some of the authors in the list appear to be usernames, but there's a list of names [here](https://www.epirhandbook.com/en/#contributors) you should be able to refer to

Done, thank you for linking to the author list.

> I'm not sure that the data.org entry really needs to be a citation; probably an inline URL where you currently cite it would suffice since there doesn't seem to be a "formal" author list or DOI or anything

I've removed the citation and instead linked Epiverse-TRACE to the website in the citation.
