### 2026-03-05 #8781 @joshwlambert
There will now be small differences between the paper in the v0.7.1 release and the current version, please let me know if another version release is needed.
