### 2026-03-05 #8781 @joshwlambert
> footnote on first page continues on second in an awkward/unclear way – I bet if you slightly increase or reduce the size of the package logo it will either bump the whole thing onto the next page or allow the footnote to fit on the same page

I can confirm that the footnote is now on one page since reducing the size of the logo.
