### 2026-03-06 #8781 @rkurchin
Looks great!

> There will now be small differences between the paper in the v0.7.1 release and the current version, please let me know if another version release is needed.

Nope, if it's just the paper and not the code, no need – thanks for checking!
