### 2026-03-06 #8781 @rkurchin
Congrats @joshwlambert and thank you to editor @fabian-s and reviewers @mponce0 and @gdenicola – JOSS couldn't run without your efforts!
