### 2026-03-06 #8781 @joshwlambert
Many thanks to @fabian-s, @mponce0, @gdenicola and @rkurchin for helping in the process of this publication, and to JOSS more widely for their work to reward open-source.
